Counterfactual Partial Dependence Effects

Our selection algorithm, select_taxa depends on estimates for counterfactual partial dependence effects across all taxa and time lags. This supports estimation of these effects for a single random data split pair.

Usage

pd_effects(
  fit,
  ts,
  w0,
  w1,
  n_sample = NULL,
  patch_len = 8,
  intervention_len = NULL
)

Arguments

fit

An object of class mbtransfer_model, as generated using the mbtransfer function. This includes trained boosting models for every taxon, stored within the @parameters slot.

ts

An object of class ts_inter containing the time-varying microbiome community, environmental interventions, and static host features data. The columns for each element of the values matrix are expected to be consecutive timepoints in an changing community. ts_from_dfs() is a helper to create these objects from data.frames with the relevant information.

w0

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

w1

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

Examples

data(sim_ts)
w0 <- cbind(sim_ts[[1]]@interventions, matrix(0, nrow = 1, ncol = 3))
w1 <- cbind(sim_ts[[1]]@interventions, matrix(1, nrow = 1, ncol = 3))
fit <- mbtransfer(sim_ts, 1, 1, nrounds = 10)
pd_effects(fit, sim_ts, w0, w1)
#>        S8_S1_T1 S8_S1_T2 S8_S1_T3 S8_S1_T4 S8_S1_T5 S8_S1_T6 S8_S1_T7 S8_S1_T8
#> tax1          0        0        0        0        0        0        0        0
#> tax2          0        0        0        0        0        0        0        0
#> tax3          0        0        0        0        0        0        0        0
#> tax4          0        0        0        0        0        0        0        0
#> tax5          0        0        0        0        0        0        0        0
#> tax6          0        0        0        0        0        0        0        0
#> tax7          0        0        0        0        0        0        0        0
#> tax8          0        0        0        0        0        0        0        0
#> tax9          0        0        0        0        0        0        0        0
#> tax10         0        0        0        0        0        0        0        0
#> tax11         0        0        0        0        0        0        0        0
#> tax12         0        0        0        0        0        0        0        0
#> tax13         0        0        0        0        0        0        0        0
#> tax14         0        0        0        0        0        0        0        0
#> tax15         0        0        0        0        0        0        0        0
#> tax16         0        0        0        0        0        0        0        0
#> tax17         0        0        0        0        0        0        0        0
#> tax18         0        0        0        0        0        0        0        0
#> tax19         0        0        0        0        0        0        0        0
#> tax20         0        0        0        0        0        0        0        0
#> tax21         0        0        0        0        0        0        0        0
#> tax22         0        0        0        0        0        0        0        0
#> tax23         0        0        0        0        0        0        0        0
#> tax24         0        0        0        0        0        0        0        0
#> tax25         0        0        0        0        0        0        0        0
#> tax26         0        0        0        0        0        0        0        0
#> tax27         0        0        0        0        0        0        0        0
#> tax28         0        0        0        0        0        0        0        0
#> tax29         0        0        0        0        0        0        0        0
#> tax30         0        0        0        0        0        0        0        0
#> tax31         0        0        0        0        0        0        0        0
#> tax32         0        0        0        0        0        0        0        0
#> tax33         0        0        0        0        0        0        0        0
#> tax34         0        0        0        0        0        0        0        0
#> tax35         0        0        0        0        0        0        0        0
#> tax36         0        0        0        0        0        0        0        0
#> tax37         0        0        0        0        0        0        0        0
#> tax38         0        0        0        0        0        0        0        0
#> tax39         0        0        0        0        0        0        0        0
#> tax40         0        0        0        0        0        0        0        0
#> tax41         0        0        0        0        0        0        0        0
#> tax42         0        0        0        0        0        0        0        0
#> tax43         0        0        0        0        0        0        0        0
#> tax44         0        0        0        0        0        0        0        0
#> tax45         0        0        0        0        0        0        0        0
#> tax46         0        0        0        0        0        0        0        0
#> tax47         0        0        0        0        0        0        0        0
#> tax48         0        0        0        0        0        0        0        0
#> tax49         0        0        0        0        0        0        0        0
#> tax50         0        0        0        0        0        0        0        0
#> tax51         0        0        0        0        0        0        0        0
#> tax52         0        0        0        0        0        0        0        0
#> tax53         0        0        0        0        0        0        0        0
#> tax54         0        0        0        0        0        0        0        0
#> tax55         0        0        0        0        0        0        0        0
#> tax56         0        0        0        0        0        0        0        0
#> tax57         0        0        0        0        0        0        0        0
#> tax58         0        0        0        0        0        0        0        0
#> tax59         0        0        0        0        0        0        0        0
#> tax60         0        0        0        0        0        0        0        0
#> tax61         0        0        0        0        0        0        0        0
#> tax62         0        0        0        0        0        0        0        0
#> tax63         0        0        0        0        0        0        0        0
#> tax64         0        0        0        0        0        0        0        0
#> tax65         0        0        0        0        0        0        0        0
#> tax66         0        0        0        0        0        0        0        0
#> tax67         0        0        0        0        0        0        0        0
#> tax68         0        0        0        0        0        0        0        0
#> tax69         0        0        0        0        0        0        0        0
#> tax70         0        0        0        0        0        0        0        0
#> tax71         0        0        0        0        0        0        0        0
#> tax72         0        0        0        0        0        0        0        0
#> tax73         0        0        0        0        0        0        0        0
#> tax74         0        0        0        0        0        0        0        0
#> tax75         0        0        0        0        0        0        0        0
#> tax76         0        0        0        0        0        0        0        0
#> tax77         0        0        0        0        0        0        0        0
#> tax78         0        0        0        0        0        0        0        0
#> tax79         0        0        0        0        0        0        0        0
#> tax80         0        0        0        0        0        0        0        0
#> tax81         0        0        0        0        0        0        0        0
#> tax82         0        0        0        0        0        0        0        0
#> tax83         0        0        0        0        0        0        0        0
#> tax84         0        0        0        0        0        0        0        0
#> tax85         0        0        0        0        0        0        0        0
#> tax86         0        0        0        0        0        0        0        0
#> tax87         0        0        0        0        0        0        0        0
#> tax88         0        0        0        0        0        0        0        0
#> tax89         0        0        0        0        0        0        0        0
#> tax90         0        0        0        0        0        0        0        0
#> tax91         0        0        0        0        0        0        0        0
#> tax92         0        0        0        0        0        0        0        0
#> tax93         0        0        0        0        0        0        0        0
#> tax94         0        0        0        0        0        0        0        0
#> tax95         0        0        0        0        0        0        0        0
#> tax96         0        0        0        0        0        0        0        0
#> tax97         0        0        0        0        0        0        0        0
#> tax98         0        0        0        0        0        0        0        0
#> tax99         0        0        0        0        0        0        0        0
#> tax100        0        0        0        0        0        0        0        0
#>        S8_S1_T9 S8_S1_T10 S8_S1_T11 S8_S1_T12 S8_S1_T13 S8_S1_T14 S8_S1_T15
#> tax1          0         0         0         0         0         0         0
#> tax2          0         0         0         0         0         0         0
#> tax3          0         0         0         0         0         0         0
#> tax4          0         0         0         0         0         0         0
#> tax5          0         0         0         0         0         0         0
#> tax6          0         0         0         0         0         0         0
#> tax7          0         0         0         0         0         0         0
#> tax8          0         0         0         0         0         0         0
#> tax9          0         0         0         0         0         0         0
#> tax10         0         0         0         0         0         0         0
#> tax11         0         0         0         0         0         0         0
#> tax12         0         0         0         0         0         0         0
#> tax13         0         0         0         0         0         0         0
#> tax14         0         0         0         0         0         0         0
#> tax15         0         0         0         0         0         0         0
#> tax16         0         0         0         0         0         0         0
#> tax17         0         0         0         0         0         0         0
#> tax18         0         0         0         0         0         0         0
#> tax19         0         0         0         0         0         0         0
#> tax20         0         0         0         0         0         0         0
#> tax21         0         0         0         0         0         0         0
#> tax22         0         0         0         0         0         0         0
#> tax23         0         0         0         0         0         0         0
#> tax24         0         0         0         0         0         0         0
#> tax25         0         0         0         0         0         0         0
#> tax26         0         0         0         0         0         0         0
#> tax27         0         0         0         0         0         0         0
#> tax28         0         0         0         0         0         0         0
#> tax29         0         0         0         0         0         0         0
#> tax30         0         0         0         0         0         0         0
#> tax31         0         0         0         0         0         0         0
#> tax32         0         0         0         0         0         0         0
#> tax33         0         0         0         0         0         0         0
#> tax34         0         0         0         0         0         0         0
#> tax35         0         0         0         0         0         0         0
#> tax36         0         0         0         0         0         0         0
#> tax37         0         0         0         0         0         0         0
#> tax38         0         0         0         0         0         0         0
#> tax39         0         0         0         0         0         0         0
#> tax40         0         0         0         0         0         0         0
#> tax41         0         0         0         0         0         0         0
#> tax42         0         0         0         0         0         0         0
#> tax43         0         0         0         0         0         0         0
#> tax44         0         0         0         0         0         0         0
#> tax45         0         0         0         0         0         0         0
#> tax46         0         0         0         0         0         0         0
#> tax47         0         0         0         0         0         0         0
#> tax48         0         0         0         0         0         0         0
#> tax49         0         0         0         0         0         0         0
#> tax50         0         0         0         0         0         0         0
#> tax51         0         0         0         0         0         0         0
#> tax52         0         0         0         0         0         0         0
#> tax53         0         0         0         0         0         0         0
#> tax54         0         0         0         0         0         0         0
#> tax55         0         0         0         0         0         0         0
#> tax56         0         0         0         0         0         0         0
#> tax57         0         0         0         0         0         0         0
#> tax58         0         0         0         0         0         0         0
#> tax59         0         0         0         0         0         0         0
#> tax60         0         0         0         0         0         0         0
#> tax61         0         0         0         0         0         0         0
#> tax62         0         0         0         0         0         0         0
#> tax63         0         0         0         0         0         0         0
#> tax64         0         0         0         0         0         0         0
#> tax65         0         0         0         0         0         0         0
#> tax66         0         0         0         0         0         0         0
#> tax67         0         0         0         0         0         0         0
#> tax68         0         0         0         0         0         0         0
#> tax69         0         0         0         0         0         0         0
#> tax70         0         0         0         0         0         0         0
#> tax71         0         0         0         0         0         0         0
#> tax72         0         0         0         0         0         0         0
#> tax73         0         0         0         0         0         0         0
#> tax74         0         0         0         0         0         0         0
#> tax75         0         0         0         0         0         0         0
#> tax76         0         0         0         0         0         0         0
#> tax77         0         0         0         0         0         0         0
#> tax78         0         0         0         0         0         0         0
#> tax79         0         0         0         0         0         0         0
#> tax80         0         0         0         0         0         0         0
#> tax81         0         0         0         0         0         0         0
#> tax82         0         0         0         0         0         0         0
#> tax83         0         0         0         0         0         0         0
#> tax84         0         0         0         0         0         0         0
#> tax85         0         0         0         0         0         0         0
#> tax86         0         0         0         0         0         0         0
#> tax87         0         0         0         0         0         0         0
#> tax88         0         0         0         0         0         0         0
#> tax89         0         0         0         0         0         0         0
#> tax90         0         0         0         0         0         0         0
#> tax91         0         0         0         0         0         0         0
#> tax92         0         0         0         0         0         0         0
#> tax93         0         0         0         0         0         0         0
#> tax94         0         0         0         0         0         0         0
#> tax95         0         0         0         0         0         0         0
#> tax96         0         0         0         0         0         0         0
#> tax97         0         0         0         0         0         0         0
#> tax98         0         0         0         0         0         0         0
#> tax99         0         0         0         0         0         0         0
#> tax100        0         0         0         0         0         0         0
#>        S8_S1_T16 S8_S1_T17 S8_S1_T18 S8_S1_T19 S8_S1_T20 S8_S1_T21 S8_S1_T22
#> tax1           0         0         0         0         0         0         0
#> tax2           0         0         0         0         0         0         0
#> tax3           0         0         0         0         0         0         0
#> tax4           0         0         0         0         0         0         0
#> tax5           0         0         0         0         0         0         0
#> tax6           0         0         0         0         0         0         0
#> tax7           0         0         0         0         0         0         0
#> tax8           0         0         0         0         0         0         0
#> tax9           0         0         0         0         0         0         0
#> tax10          0         0         0         0         0         0         0
#> tax11          0         0         0         0         0         0         0
#> tax12          0         0         0         0         0         0         0
#> tax13          0         0         0         0         0         0         0
#> tax14          0         0         0         0         0         0         0
#> tax15          0         0         0         0         0         0         0
#> tax16          0         0         0         0         0         0         0
#> tax17          0         0         0         0         0         0         0
#> tax18          0         0         0         0         0         0         0
#> tax19          0         0         0         0         0         0         0
#> tax20          0         0         0         0         0         0         0
#> tax21          0         0         0         0         0         0         0
#> tax22          0         0         0         0         0         0         0
#> tax23          0         0         0         0         0         0         0
#> tax24          0         0         0         0         0         0         0
#> tax25          0         0         0         0         0         0         0
#> tax26          0         0         0         0         0         0         0
#> tax27          0         0         0         0         0         0         0
#> tax28          0         0         0         0         0         0         0
#> tax29          0         0         0         0         0         0         0
#> tax30          0         0         0         0         0         0         0
#> tax31          0         0         0         0         0         0         0
#> tax32          0         0         0         0         0         0         0
#> tax33          0         0         0         0         0         0         0
#> tax34          0         0         0         0         0         0         0
#> tax35          0         0         0         0         0         0         0
#> tax36          0         0         0         0         0         0         0
#> tax37          0         0         0         0         0         0         0
#> tax38          0         0         0         0         0         0         0
#> tax39          0         0         0         0         0         0         0
#> tax40          0         0         0         0         0         0         0
#> tax41          0         0         0         0         0         0         0
#> tax42          0         0         0         0         0         0         0
#> tax43          0         0         0         0         0         0         0
#> tax44          0         0         0         0         0         0         0
#> tax45          0         0         0         0         0         0         0
#> tax46          0         0         0         0         0         0         0
#> tax47          0         0         0         0         0         0         0
#> tax48          0         0         0         0         0         0         0
#> tax49          0         0         0         0         0         0         0
#> tax50          0         0         0         0         0         0         0
#> tax51          0         0         0         0         0         0         0
#> tax52          0         0         0         0         0         0         0
#> tax53          0         0         0         0         0         0         0
#> tax54          0         0         0         0         0         0         0
#> tax55          0         0         0         0         0         0         0
#> tax56          0         0         0         0         0         0         0
#> tax57          0         0         0         0         0         0         0
#> tax58          0         0         0         0         0         0         0
#> tax59          0         0         0         0         0         0         0
#> tax60          0         0         0         0         0         0         0
#> tax61          0         0         0         0         0         0         0
#> tax62          0         0         0         0         0         0         0
#> tax63          0         0         0         0         0         0         0
#> tax64          0         0         0         0         0         0         0
#> tax65          0         0         0         0         0         0         0
#> tax66          0         0         0         0         0         0         0
#> tax67          0         0         0         0         0         0         0
#> tax68          0         0         0         0         0         0         0
#> tax69          0         0         0         0         0         0         0
#> tax70          0         0         0         0         0         0         0
#> tax71          0         0         0         0         0         0         0
#> tax72          0         0         0         0         0         0         0
#> tax73          0         0         0         0         0         0         0
#> tax74          0         0         0         0         0         0         0
#> tax75          0         0         0         0         0         0         0
#> tax76          0         0         0         0         0         0         0
#> tax77          0         0         0         0         0         0         0
#> tax78          0         0         0         0         0         0         0
#> tax79          0         0         0         0         0         0         0
#> tax80          0         0         0         0         0         0         0
#> tax81          0         0         0         0         0         0         0
#> tax82          0         0         0         0         0         0         0
#> tax83          0         0         0         0         0         0         0
#> tax84          0         0         0         0         0         0         0
#> tax85          0         0         0         0         0         0         0
#> tax86          0         0         0         0         0         0         0
#> tax87          0         0         0         0         0         0         0
#> tax88          0         0         0         0         0         0         0
#> tax89          0         0         0         0         0         0         0
#> tax90          0         0         0         0         0         0         0
#> tax91          0         0         0         0         0         0         0
#> tax92          0         0         0         0         0         0         0
#> tax93          0         0         0         0         0         0         0
#> tax94          0         0         0         0         0         0         0
#> tax95          0         0         0         0         0         0         0
#> tax96          0         0         0         0         0         0         0
#> tax97          0         0         0         0         0         0         0
#> tax98          0         0         0         0         0         0         0
#> tax99          0         0         0         0         0         0         0
#> tax100         0         0         0         0         0         0         0
#>        S8_S1_T23 S8_S1_T24 S8_S1_T25 S8_S1_T26 S8_S1_T27           S8_
#> tax1           0         0         0         0         0 -0.6327717134
#> tax2           0         0         0         0         0 -8.9875582513
#> tax3           0         0         0         0         0  0.3937873840
#> tax4           0         0         0         0         0 -0.0201892853
#> tax5           0         0         0         0         0 -4.4866585618
#> tax6           0         0         0         0         0 -0.8582592010
#> tax7           0         0         0         0         0 -2.1543407440
#> tax8           0         0         0         0         0 -3.1240035693
#> tax9           0         0         0         0         0 -2.2802833603
#> tax10          0         0         0         0         0 -4.6892160802
#> tax11          0         0         0         0         0  0.0034351349
#> tax12          0         0         0         0         0 -0.0204715729
#> tax13          0         0         0         0         0  0.1654682159
#> tax14          0         0         0         0         0 -0.0492248535
#> tax15          0         0         0         0         0 -0.0553627014
#> tax16          0         0         0         0         0 -0.2127256393
#> tax17          0         0         0         0         0  0.0907707214
#> tax18          0         0         0         0         0  0.1127681505
#> tax19          0         0         0         0         0  0.3558053970
#> tax20          0         0         0         0         0  0.5128345490
#> tax21          0         0         0         0         0  0.0248193741
#> tax22          0         0         0         0         0 -0.1303119659
#> tax23          0         0         0         0         0  0.0995483398
#> tax24          0         0         0         0         0 -0.0324039459
#> tax25          0         0         0         0         0  0.2163619995
#> tax26          0         0         0         0         0 -0.1319246292
#> tax27          0         0         0         0         0 -0.1229515076
#> tax28          0         0         0         0         0  0.0561218262
#> tax29          0         0         0         0         0  0.3455181122
#> tax30          0         0         0         0         0 -0.0793304443
#> tax31          0         0         0         0         0  0.1605405807
#> tax32          0         0         0         0         0  0.1207734517
#> tax33          0         0         0         0         0  0.0938978195
#> tax34          0         0         0         0         0  0.0489125252
#> tax35          0         0         0         0         0  0.0000000000
#> tax36          0         0         0         0         0 -0.1186246872
#> tax37          0         0         0         0         0  0.0007133484
#> tax38          0         0         0         0         0 -0.0740737915
#> tax39          0         0         0         0         0  0.0066747665
#> tax40          0         0         0         0         0  0.1884799060
#> tax41          0         0         0         0         0 -0.0001268387
#> tax42          0         0         0         0         0  0.2313776016
#> tax43          0         0         0         0         0 -0.2732028961
#> tax44          0         0         0         0         0  0.0319805145
#> tax45          0         0         0         0         0  0.1965379715
#> tax46          0         0         0         0         0  0.1439895630
#> tax47          0         0         0         0         0  0.0422534943
#> tax48          0         0         0         0         0 -0.1410312653
#> tax49          0         0         0         0         0 -0.0736150742
#> tax50          0         0         0         0         0  0.5354537964
#> tax51          0         0         0         0         0 -0.1536760330
#> tax52          0         0         0         0         0  0.1007843018
#> tax53          0         0         0         0         0 -0.1478652954
#> tax54          0         0         0         0         0 -0.3852119446
#> tax55          0         0         0         0         0  0.0322227478
#> tax56          0         0         0         0         0  0.1188507080
#> tax57          0         0         0         0         0  0.2763490677
#> tax58          0         0         0         0         0 -0.0009775162
#> tax59          0         0         0         0         0 -0.1131505966
#> tax60          0         0         0         0         0  0.0277347565
#> tax61          0         0         0         0         0 -0.0244722366
#> tax62          0         0         0         0         0 -0.0082368851
#> tax63          0         0         0         0         0 -0.0846949986
#> tax64          0         0         0         0         0 -0.0102405548
#> tax65          0         0         0         0         0  0.2671346664
#> tax66          0         0         0         0         0  0.0753459930
#> tax67          0         0         0         0         0 -0.4326419830
#> tax68          0         0         0         0         0 -0.0620737076
#> tax69          0         0         0         0         0  0.1088876724
#> tax70          0         0         0         0         0  0.0013914108
#> tax71          0         0         0         0         0 -0.0693349838
#> tax72          0         0         0         0         0 -0.4036769867
#> tax73          0         0         0         0         0  0.1814718246
#> tax74          0         0         0         0         0  0.0478010178
#> tax75          0         0         0         0         0  0.1415758133
#> tax76          0         0         0         0         0  0.0626258850
#> tax77          0         0         0         0         0 -0.0986537933
#> tax78          0         0         0         0         0  0.0439052582
#> tax79          0         0         0         0         0 -0.3178339005
#> tax80          0         0         0         0         0 -0.0820455551
#> tax81          0         0         0         0         0  0.0042743683
#> tax82          0         0         0         0         0 -0.0115900040
#> tax83          0         0         0         0         0 -0.3128871918
#> tax84          0         0         0         0         0  0.0110177994
#> tax85          0         0         0         0         0 -0.2760499205
#> tax86          0         0         0         0         0  0.1508111954
#> tax87          0         0         0         0         0  0.1465444565
#> tax88          0         0         0         0         0  0.2416343689
#> tax89          0         0         0         0         0  0.0470733643
#> tax90          0         0         0         0         0 -0.0028324127
#> tax91          0         0         0         0         0 -0.3564462662
#> tax92          0         0         0         0         0  0.1078033447
#> tax93          0         0         0         0         0 -0.0472717285
#> tax94          0         0         0         0         0  0.0902631169
#> tax95          0         0         0         0         0 -0.2134904861
#> tax96          0         0         0         0         0 -0.0975799561
#> tax97          0         0         0         0         0  0.0443878174
#> tax98          0         0         0         0         0  0.0540866852
#> tax99          0         0         0         0         0  0.3174533844
#> tax100         0         0         0         0         0  0.0797128677
#>                  S8_           S8_
#> tax1   -0.8478096042 -0.8675073328
#> tax2   -8.7428761891 -8.7773564430
#> tax3    0.5063074089  0.5237090247
#> tax4   -0.3665837787 -0.3849877630
#> tax5   -5.2434905370 -5.2735959462
#> tax6   -1.1589495114 -1.1916984263
#> tax7   -1.9992012410 -1.9819612390
#> tax8   -3.2874353613 -3.3004115423
#> tax9   -1.8929547923 -1.8838966120
#> tax10  -4.7895319462 -4.8012747992
#> tax11   0.0460575308  0.0483730521
#> tax12   0.0598706972  0.0852133887
#> tax13   0.1633217335  0.1681732904
#> tax14  -0.0089849631 -0.0011290482
#> tax15  -0.0877813612 -0.0975409008
#> tax16  -0.2040239402 -0.2030890442
#> tax17   0.2012204897  0.1937590667
#> tax18   0.0995545841  0.0736776874
#> tax19   0.3403929188  0.3523161525
#> tax20   0.5566146260  0.5481644017
#> tax21   0.0190494288  0.0103687672
#> tax22  -0.3314572402 -0.3424074536
#> tax23   0.1381494431  0.1504232997
#> tax24   0.0565847783  0.0441313471
#> tax25   0.2391087101  0.2626871041
#> tax26  -0.2623428163 -0.2859090283
#> tax27  -0.0821597122 -0.0776120935
#> tax28  -0.1465803896 -0.1625125408
#> tax29   0.4616309121  0.4748309226
#> tax30  -0.0356994129 -0.0368100462
#> tax31   0.1178859529  0.1439979417
#> tax32   0.1581298056  0.1622465338
#> tax33   0.2621377877  0.2655755111
#> tax34   0.0283841519  0.0200348922
#> tax35  -0.0033049924 -0.0131999311
#> tax36  -0.1074204445 -0.1093813238
#> tax37   0.0179830279  0.0491062573
#> tax38  -0.0212670848 -0.0169758910
#> tax39  -0.0504596233 -0.0372670037
#> tax40   0.2294097741  0.2336868218
#> tax41   0.0337354683  0.0185982613
#> tax42   0.2719012556  0.2628679389
#> tax43  -0.3354143415 -0.3246603580
#> tax44  -0.0277935323 -0.0405326457
#> tax45   0.2005165077  0.2218562421
#> tax46   0.0338571526  0.0312566530
#> tax47   0.0233924275  0.0092149803
#> tax48  -0.1214756739 -0.1148470470
#> tax49   0.1014556771  0.1129363037
#> tax50   0.5423200585  0.5302766845
#> tax51  -0.1389137279 -0.1612966855
#> tax52   0.0988244556  0.1010853904
#> tax53  -0.0333259333 -0.0313971270
#> tax54  -0.4700585433 -0.4606595607
#> tax55   0.0261057899  0.0001327651
#> tax56   0.1547277882  0.1500026499
#> tax57   0.3723915759  0.3954967544
#> tax58  -0.0002958434 -0.0074470270
#> tax59  -0.0986129443 -0.0949361892
#> tax60   0.0718001298  0.0825833820
#> tax61   0.2196457954  0.2651271820
#> tax62   0.0484756175  0.0660507793
#> tax63  -0.1081907919 -0.1136227222
#> tax64  -0.1240350746 -0.1547520274
#> tax65   0.0838585808  0.0699041003
#> tax66   0.0724313940  0.0707149846
#> tax67  -0.4573414666 -0.4651363009
#> tax68  -0.1389463175 -0.1478569735
#> tax69   0.2070452259  0.2178796814
#> tax70   0.0280236630  0.0345140752
#> tax71  -0.0755662805 -0.0790973618
#> tax72  -0.5561947596 -0.5602233410
#> tax73   0.1488138381  0.1479273410
#> tax74  -0.1134141513 -0.1274025100
#> tax75   0.2153189182  0.2164079235
#> tax76   0.0248673303  0.0147345747
#> tax77  -0.1510234787 -0.1584770452
#> tax78   0.0533345200  0.0628850233
#> tax79  -0.3273007302 -0.3322032406
#> tax80  -0.1243201097 -0.1553293864
#> tax81   0.0747567586  0.1019208318
#> tax82  -0.0541457222 -0.0400643235
#> tax83  -0.2753307706 -0.2835043839
#> tax84   0.0192914463  0.0103514876
#> tax85  -0.2932260718 -0.2905181703
#> tax86   0.0916868846  0.0917179925
#> tax87   0.1119051093  0.1084583600
#> tax88   0.2764215583  0.2676472437
#> tax89   0.0175249009  0.0141111556
#> tax90  -0.0076473213 -0.0093630836
#> tax91  -0.4163475264 -0.4132453146
#> tax92   0.2276004723  0.2230035237
#> tax93  -0.0846238136 -0.0732455367
#> tax94   0.2385282971  0.2438651494
#> tax95  -0.1824761345 -0.1817482426
#> tax96  -0.0419797330 -0.0362776121
#> tax97   0.3034799667  0.3146126384
#> tax98   0.0798123564  0.0680286884
#> tax99   0.2045073736  0.1968304430
#> tax100  0.1255884511  0.1268000262