Mirror Splits from Partial Dependence

For ech split, we return an array of dimension n_taxa x 2 fits x n_lags. Each entry contains the estimated counterfactual effect for that taxon and lag combination across the two models (each fit on a different random sample of data).

Usage

pd_splits(ts, w0, w1, tr_fun, n_splits = 20, ...)

Arguments

ts

An object of class ts_inter containing the time-varying microbiome community, environmental interventions, and static host features data. The columns for each element of the values matrix are expected to be consecutive timepoints in an changing community. ts_from_dfs() is a helper to create these objects from data.frames with the relevant information.

w0

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

w1

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

tr_fun

A function that can be used to train over random splits. In the examples in this package, we use an anonymous function that fills our chosen hyperparameters in mbtransfer. For example \(x) mbtransfer(x, P = 1, Q = 1) will fit a lag-1 transfer function model on all the random splits.

Examples

library(mbtransfer)
data(sim_ts)
w0 <- cbind(sim_ts[[1]]@interventions, matrix(0, nrow = 1, ncol = 3))
w1 <- cbind(sim_ts[[1]]@interventions, matrix(1, nrow = 1, ncol = 3))
pd_splits(sim_ts, w0, w1, \(x) mbtransfer(x, 1, 1, nrounds = 10), n_splits = 1)
#> Training models for split 1/1
#> [[1]]
#> , , 1
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 2
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 3
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 4
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 5
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 6
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 7
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 8
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 9
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 10
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 11
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 12
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 13
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 14
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 15
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 16
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 17
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 18
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 19
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 20
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 21
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 22
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 23
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 24
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 25
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 26
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 27
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 28
#> 
#>                [,1]         [,2]
#>  [1,] -2.4805431366 -3.002172470
#>  [2,]  3.1570493834  3.385816495
#>  [3,] -1.7414741516 -1.157933235
#>  [4,] -9.6943910917 -9.758343401
#>  [5,]  3.2642822266  3.199480443
#>  [6,]  1.0047260239  1.204376470
#>  [7,] -1.7230110168 -1.815294266
#>  [8,]  0.1020622253  0.699337959
#>  [9,] -0.0102109909 -0.096532822
#> [10,] -0.0962247849 -0.217235656
#> [11,]  0.2647193159  0.099644195
#> [12,] -0.3701667786  0.135799408
#> [13,] -0.0126047134 -0.219240189
#> [14,] -0.3698231152 -0.211420195
#> [15,] -0.4579296112  0.177540779
#> [16,] -0.0185452302  0.440642311
#> [17,] -0.4131069183 -0.236657165
#> [18,] -0.0230731964 -0.444272995
#> [19,] -0.3594017029 -0.343375706
#> [20,]  0.0920038223  0.280948650
#> [21,] -0.0039701462  0.012681007
#> [22,] -1.0499773026 -0.788928986
#> [23,]  0.6175890196  0.180898666
#> [24,] -0.0131320953  0.274586678
#> [25,]  0.1541595459  0.066898346
#> [26,] -0.2746791840  0.186372757
#> [27,] -0.0419206619  0.327350401
#> [28,]  0.0743951797 -0.271641731
#> [29,] -0.0299396515 -0.003025055
#> [30,] -0.4889621735 -0.668946266
#> [31,] -0.7620868683 -0.508547760
#> [32,] -0.3189888000  0.088536263
#> [33,]  0.2982025146  0.238865807
#> [34,] -0.0086669922  0.528608333
#> [35,] -0.1837539673 -0.353092228
#> [36,] -1.8979530334 -1.132450104
#> [37,]  0.0006504059 -0.317517303
#> [38,] -0.4331769943 -0.729449749
#> [39,] -0.2868537903  0.311318398
#> [40,] -0.1757287979 -0.290233612
#> [41,] -0.7980918884  0.381481171
#> [42,] -0.7494115602 -1.250482559
#> [43,]  0.0936622620  0.054176331
#> [44,] -0.1314888000  0.002673172
#> [45,] -0.7866668701 -0.262976646
#> [46,]  0.3028135300  0.065480210
#> [47,]  0.3862638474  0.569713593
#> [48,]  0.1493835449 -0.508293152
#> [49,] -0.0997915268 -0.006019592
#> [50,] -0.4501457441 -0.763053326
#> 
#> , , 29
#> 
#>               [,1]          [,2]
#>  [1,]  -3.35308045  -4.080601579
#>  [2,]   3.64117555   4.073966810
#>  [3,]  -2.29124428  -2.038988772
#>  [4,] -10.66380019 -10.937593869
#>  [5,]   3.35643843   3.393440655
#>  [6,]   0.98538901   1.191161281
#>  [7,]  -2.25022529  -2.445356028
#>  [8,]   0.12229222   0.567847252
#>  [9,]  -0.02720912  -0.113917782
#> [10,]  -0.22731225  -0.283448106
#> [11,]   0.12511893  -0.049686795
#> [12,]  -0.72449709  -0.208539985
#> [13,]  -0.10435872  -0.506678014
#> [14,]  -0.49119302  -0.265186537
#> [15,]  -0.37281713   0.178392637
#> [16,]   0.17188373   0.803149621
#> [17,]  -0.34694617  -0.135627474
#> [18,]  -0.37578519  -0.700425375
#> [19,]  -0.89896243  -0.867457685
#> [20,]   0.06133261   0.221427316
#> [21,]  -0.05512158   0.065371468
#> [22,]  -1.11214763  -0.990527857
#> [23,]   0.36847959   0.425475461
#> [24,]  -0.30663565   0.009207317
#> [25,]  -0.13449351  -0.170756204
#> [26,]  -0.27136303  -0.042619887
#> [27,]  -0.37410710   0.407018071
#> [28,]  -0.04822888  -0.405386698
#> [29,]  -0.11880486  -0.173510086
#> [30,]  -0.67189382  -0.519684247
#> [31,]  -1.04044978  -0.629167863
#> [32,]  -0.11050556   0.023765950
#> [33,]   0.25980895   0.064498084
#> [34,]  -0.08010921   0.522702524
#> [35,]  -0.35997218  -0.401303212
#> [36,]  -2.25688560  -1.688724541
#> [37,]  -0.23120960  -0.615915798
#> [38,]  -0.60209395  -0.723787262
#> [39,]  -0.30591413   0.200891813
#> [40,]  -0.56332048  -0.935299964
#> [41,]  -0.76907335   0.345545428
#> [42,]  -1.39240056  -1.923549607
#> [43,]   0.17491747   0.042811121
#> [44,]  -0.16182268  -0.144237791
#> [45,]  -1.41109489  -0.778589498
#> [46,]   0.23428433   0.216137023
#> [47,]   0.13405539   0.483172053
#> [48,]   0.18621345  -0.556046940
#> [49,]  -0.16293094  -0.028629712
#> [50,]  -0.96935278  -1.840239752
#> 
#> , , 30
#> 
#>                [,1]          [,2]
#>  [1,]  -3.596084481 -4.437087e+00
#>  [2,]   3.624835616  4.094123e+00
#>  [3,]  -2.535591670 -2.351396e+00
#>  [4,] -11.007157417 -1.134844e+01
#>  [5,]   3.212571235  3.306581e+00
#>  [6,]   0.889309520  1.075257e+00
#>  [7,]  -2.451796214 -2.716570e+00
#>  [8,]   0.012889930  3.659635e-01
#>  [9,]  -0.175687654 -2.098667e-01
#> [10,]  -0.394998460 -3.946228e-01
#> [11,]   0.004408814 -1.706550e-01
#> [12,]  -0.931250867 -3.820853e-01
#> [13,]  -0.243038881 -6.805670e-01
#> [14,]  -0.618538925 -3.411142e-01
#> [15,]  -0.502897058  6.113611e-02
#> [16,]   0.139266639  7.644183e-01
#> [17,]  -0.459872609 -2.351317e-01
#> [18,]  -0.571731636 -8.991374e-01
#> [19,]  -1.117373648 -1.067682e+00
#> [20,]  -0.037399542  9.535237e-02
#> [21,]  -0.171340511 -3.688903e-04
#> [22,]  -1.220786481 -1.196192e+00
#> [23,]   0.167620977  3.897898e-01
#> [24,]  -0.492554165 -1.582125e-01
#> [25,]  -0.315627711 -3.181166e-01
#> [26,]  -0.409521625 -1.763560e-01
#> [27,]  -0.501780623  3.501139e-01
#> [28,]  -0.166744800 -5.289580e-01
#> [29,]  -0.186951853 -2.604589e-01
#> [30,]  -0.803690388 -6.353284e-01
#> [31,]  -1.188681375 -7.770392e-01
#> [32,]  -0.201357069 -5.610947e-02
#> [33,]   0.118640082 -1.039817e-01
#> [34,]  -0.188815344  3.910481e-01
#> [35,]  -0.475422893 -5.153139e-01
#> [36,]  -2.436655885 -1.925621e+00
#> [37,]  -0.388315405 -8.318973e-01
#> [38,]  -0.752718630 -8.106207e-01
#> [39,]  -0.439571063 -6.945542e-03
#> [40,]  -0.807266281 -1.245714e+00
#> [41,]  -0.914474124  1.421182e-01
#> [42,]  -1.590265274 -2.167270e+00
#> [43,]   0.078845683 -8.095843e-02
#> [44,]  -0.290012155 -2.457113e-01
#> [45,]  -1.647614729 -1.007838e+00
#> [46,]   0.095450560  9.978712e-02
#> [47,]  -0.017768224  3.054428e-01
#> [48,]   0.062416735 -6.521902e-01
#> [49,]  -0.276269277 -1.531884e-01
#> [50,]  -1.156399534 -2.138754e+00
#> 
#>