Generate a Predictor Matrix

patchify_df is used to generate training data. At test time, we don't need all the patches, but we do need to construct regressors for the immediate past. This will be matched to the model's original training data and used to generate predictions.

Usage

predictors(ts_inter, lags, subject)

Arguments

ts_inter

An object of class ts_inter_single over which to generate sliding windows.

lags

A vector specifying P and Q in the trained mbtransfer model.

subject

A static data frame of subject-level variables. This will be concatenated to time-varying intervention and taxonomic covariates when making predictions. This is analogous to the training process.

Examples

library(mbtransfer)
data(sim_ts)
ts <- subset_values(sim_ts, 1:5)
predictors(ts[[1]], c(2, 2), NULL)
#>      taxon1_lag2 taxon2_lag2 taxon3_lag2 taxon4_lag2 taxon5_lag2 taxon6_lag2
#> [1,]    7.964213    13.65294    27.30587    26.16813    6.826468    6.826468
#>      taxon7_lag2 taxon8_lag2 taxon9_lag2 taxon10_lag2 taxon11_lag2 taxon12_lag2
#> [1,]    11.37745    5.688723    38.68332            0     3.413234     13.65294
#>      taxon13_lag2 taxon14_lag2 taxon15_lag2 taxon16_lag2 taxon17_lag2
#> [1,]     22.75489     6.826468     5.688723            0     13.65294
#>      taxon18_lag2 taxon19_lag2 taxon20_lag2 taxon21_lag2 taxon22_lag2
#> [1,]     17.06617     4.550979     4.550979     5.688723     15.92843
#>      taxon23_lag2 taxon24_lag2 taxon25_lag2 taxon26_lag2 taxon27_lag2
#> [1,]     4.550979     6.826468     5.688723      10.2397     5.688723
#>      taxon28_lag2 taxon29_lag2 taxon30_lag2 taxon31_lag2 taxon32_lag2
#> [1,]     19.34166     3.413234     15.92843     18.20391     3.413234
#>      taxon33_lag2 taxon34_lag2 taxon35_lag2 taxon36_lag2 taxon37_lag2
#> [1,]     19.34166     22.75489     2.275489     7.964213     13.65294
#>      taxon38_lag2 taxon39_lag2 taxon40_lag2 taxon41_lag2 taxon42_lag2
#> [1,]     11.37745            0     7.964213     4.550979     21.61715
#>      taxon43_lag2 taxon44_lag2 taxon45_lag2 taxon46_lag2 taxon47_lag2
#> [1,]     25.03038     26.16813     6.826468            0     9.101957
#>      taxon48_lag2 taxon49_lag2 taxon50_lag2 taxon1_lag1 taxon2_lag1 taxon3_lag1
#> [1,]     6.826468     13.65294     4.550979    5.448472    4.358777    3.269083
#>      taxon4_lag1 taxon5_lag1 taxon6_lag1 taxon7_lag1 taxon8_lag1 taxon9_lag1
#> [1,]    4.358777    4.358777    10.89694    8.717555    11.98664     18.5248
#>      taxon10_lag1 taxon11_lag1 taxon12_lag1 taxon13_lag1 taxon14_lag1
#> [1,]     3.269083     3.269083     6.538166     9.807249     20.70419
#>      taxon15_lag1 taxon16_lag1 taxon17_lag1 taxon18_lag1 taxon19_lag1
#> [1,]     15.25572     14.16603      19.6145     7.627861     13.07633
#>      taxon20_lag1 taxon21_lag1 taxon22_lag1 taxon23_lag1 taxon24_lag1
#> [1,]     11.98664     5.448472     17.43511     32.69083     4.358777
#>      taxon25_lag1 taxon26_lag1 taxon27_lag1 taxon28_lag1 taxon29_lag1
#> [1,]     7.627861     4.358777     1.089694     11.98664     9.807249
#>      taxon30_lag1 taxon31_lag1 taxon32_lag1 taxon33_lag1 taxon34_lag1
#> [1,]     11.98664     6.538166     23.97328     15.25572     5.448472
#>      taxon35_lag1 taxon36_lag1 taxon37_lag1 taxon38_lag1 taxon39_lag1
#> [1,]     4.358777     5.448472     20.70419     10.89694     13.07633
#>      taxon40_lag1 taxon41_lag1 taxon42_lag1 taxon43_lag1 taxon44_lag1
#> [1,]     2.179389     3.269083     9.807249     5.448472     5.448472
#>      taxon45_lag1 taxon46_lag1 taxon47_lag1 taxon48_lag1 taxon49_lag1
#> [1,]     9.807249     2.179389     14.16603     15.25572      19.6145
#>      taxon50_lag1 intervention1_lag1 intervention1_lag0
#> [1,]     1.089694                  0                  0