Counterfactual Partial Dependence Effects

Our selection algorithm, select_taxa depends on estimates for counterfactual partial dependence effects across all taxa and time lags. This supports estimation of these effects for a single random data split pair.

Usage

pd_effects(
  fit,
  ts,
  w0,
  w1,
  n_sample = NULL,
  patch_len = 8,
  intervention_len = NULL
)

Arguments

fit

An object of class mbtransfer_model, as generated using the mbtransfer function. This includes trained boosting models for every taxon, stored within the @parameters slot.

ts

An object of class ts_inter containing the time-varying microbiome community, environmental interventions, and static host features data. The columns for each element of the values matrix are expected to be consecutive timepoints in an changing community. ts_from_dfs() is a helper to create these objects from data.frames with the relevant information.

w0

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

w1

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

Examples

data(sim_ts)
w0 <- cbind(sim_ts[[1]]@interventions, matrix(0, nrow = 1, ncol = 3))
w1 <- cbind(sim_ts[[1]]@interventions, matrix(1, nrow = 1, ncol = 3))
fit <- mbtransfer(sim_ts, 1, 1, nrounds = 10)
pd_effects(fit, sim_ts, w0, w1)
#>        S39_S1_T1 S39_S1_T2 S39_S1_T3 S39_S1_T4 S39_S1_T5 S39_S1_T6 S39_S1_T7
#> tax1           0         0         0         0         0         0         0
#> tax2           0         0         0         0         0         0         0
#> tax3           0         0         0         0         0         0         0
#> tax4           0         0         0         0         0         0         0
#> tax5           0         0         0         0         0         0         0
#> tax6           0         0         0         0         0         0         0
#> tax7           0         0         0         0         0         0         0
#> tax8           0         0         0         0         0         0         0
#> tax9           0         0         0         0         0         0         0
#> tax10          0         0         0         0         0         0         0
#> tax11          0         0         0         0         0         0         0
#> tax12          0         0         0         0         0         0         0
#> tax13          0         0         0         0         0         0         0
#> tax14          0         0         0         0         0         0         0
#> tax15          0         0         0         0         0         0         0
#> tax16          0         0         0         0         0         0         0
#> tax17          0         0         0         0         0         0         0
#> tax18          0         0         0         0         0         0         0
#> tax19          0         0         0         0         0         0         0
#> tax20          0         0         0         0         0         0         0
#> tax21          0         0         0         0         0         0         0
#> tax22          0         0         0         0         0         0         0
#> tax23          0         0         0         0         0         0         0
#> tax24          0         0         0         0         0         0         0
#> tax25          0         0         0         0         0         0         0
#> tax26          0         0         0         0         0         0         0
#> tax27          0         0         0         0         0         0         0
#> tax28          0         0         0         0         0         0         0
#> tax29          0         0         0         0         0         0         0
#> tax30          0         0         0         0         0         0         0
#> tax31          0         0         0         0         0         0         0
#> tax32          0         0         0         0         0         0         0
#> tax33          0         0         0         0         0         0         0
#> tax34          0         0         0         0         0         0         0
#> tax35          0         0         0         0         0         0         0
#> tax36          0         0         0         0         0         0         0
#> tax37          0         0         0         0         0         0         0
#> tax38          0         0         0         0         0         0         0
#> tax39          0         0         0         0         0         0         0
#> tax40          0         0         0         0         0         0         0
#> tax41          0         0         0         0         0         0         0
#> tax42          0         0         0         0         0         0         0
#> tax43          0         0         0         0         0         0         0
#> tax44          0         0         0         0         0         0         0
#> tax45          0         0         0         0         0         0         0
#> tax46          0         0         0         0         0         0         0
#> tax47          0         0         0         0         0         0         0
#> tax48          0         0         0         0         0         0         0
#> tax49          0         0         0         0         0         0         0
#> tax50          0         0         0         0         0         0         0
#> tax51          0         0         0         0         0         0         0
#> tax52          0         0         0         0         0         0         0
#> tax53          0         0         0         0         0         0         0
#> tax54          0         0         0         0         0         0         0
#> tax55          0         0         0         0         0         0         0
#> tax56          0         0         0         0         0         0         0
#> tax57          0         0         0         0         0         0         0
#> tax58          0         0         0         0         0         0         0
#> tax59          0         0         0         0         0         0         0
#> tax60          0         0         0         0         0         0         0
#> tax61          0         0         0         0         0         0         0
#> tax62          0         0         0         0         0         0         0
#> tax63          0         0         0         0         0         0         0
#> tax64          0         0         0         0         0         0         0
#> tax65          0         0         0         0         0         0         0
#> tax66          0         0         0         0         0         0         0
#> tax67          0         0         0         0         0         0         0
#> tax68          0         0         0         0         0         0         0
#> tax69          0         0         0         0         0         0         0
#> tax70          0         0         0         0         0         0         0
#> tax71          0         0         0         0         0         0         0
#> tax72          0         0         0         0         0         0         0
#> tax73          0         0         0         0         0         0         0
#> tax74          0         0         0         0         0         0         0
#> tax75          0         0         0         0         0         0         0
#> tax76          0         0         0         0         0         0         0
#> tax77          0         0         0         0         0         0         0
#> tax78          0         0         0         0         0         0         0
#> tax79          0         0         0         0         0         0         0
#> tax80          0         0         0         0         0         0         0
#> tax81          0         0         0         0         0         0         0
#> tax82          0         0         0         0         0         0         0
#> tax83          0         0         0         0         0         0         0
#> tax84          0         0         0         0         0         0         0
#> tax85          0         0         0         0         0         0         0
#> tax86          0         0         0         0         0         0         0
#> tax87          0         0         0         0         0         0         0
#> tax88          0         0         0         0         0         0         0
#> tax89          0         0         0         0         0         0         0
#> tax90          0         0         0         0         0         0         0
#> tax91          0         0         0         0         0         0         0
#> tax92          0         0         0         0         0         0         0
#> tax93          0         0         0         0         0         0         0
#> tax94          0         0         0         0         0         0         0
#> tax95          0         0         0         0         0         0         0
#> tax96          0         0         0         0         0         0         0
#> tax97          0         0         0         0         0         0         0
#> tax98          0         0         0         0         0         0         0
#> tax99          0         0         0         0         0         0         0
#> tax100         0         0         0         0         0         0         0
#>        S39_S1_T8 S39_S1_T9 S39_S1_T10 S39_S1_T11 S39_S1_T12 S39_S1_T13
#> tax1           0         0          0          0          0          0
#> tax2           0         0          0          0          0          0
#> tax3           0         0          0          0          0          0
#> tax4           0         0          0          0          0          0
#> tax5           0         0          0          0          0          0
#> tax6           0         0          0          0          0          0
#> tax7           0         0          0          0          0          0
#> tax8           0         0          0          0          0          0
#> tax9           0         0          0          0          0          0
#> tax10          0         0          0          0          0          0
#> tax11          0         0          0          0          0          0
#> tax12          0         0          0          0          0          0
#> tax13          0         0          0          0          0          0
#> tax14          0         0          0          0          0          0
#> tax15          0         0          0          0          0          0
#> tax16          0         0          0          0          0          0
#> tax17          0         0          0          0          0          0
#> tax18          0         0          0          0          0          0
#> tax19          0         0          0          0          0          0
#> tax20          0         0          0          0          0          0
#> tax21          0         0          0          0          0          0
#> tax22          0         0          0          0          0          0
#> tax23          0         0          0          0          0          0
#> tax24          0         0          0          0          0          0
#> tax25          0         0          0          0          0          0
#> tax26          0         0          0          0          0          0
#> tax27          0         0          0          0          0          0
#> tax28          0         0          0          0          0          0
#> tax29          0         0          0          0          0          0
#> tax30          0         0          0          0          0          0
#> tax31          0         0          0          0          0          0
#> tax32          0         0          0          0          0          0
#> tax33          0         0          0          0          0          0
#> tax34          0         0          0          0          0          0
#> tax35          0         0          0          0          0          0
#> tax36          0         0          0          0          0          0
#> tax37          0         0          0          0          0          0
#> tax38          0         0          0          0          0          0
#> tax39          0         0          0          0          0          0
#> tax40          0         0          0          0          0          0
#> tax41          0         0          0          0          0          0
#> tax42          0         0          0          0          0          0
#> tax43          0         0          0          0          0          0
#> tax44          0         0          0          0          0          0
#> tax45          0         0          0          0          0          0
#> tax46          0         0          0          0          0          0
#> tax47          0         0          0          0          0          0
#> tax48          0         0          0          0          0          0
#> tax49          0         0          0          0          0          0
#> tax50          0         0          0          0          0          0
#> tax51          0         0          0          0          0          0
#> tax52          0         0          0          0          0          0
#> tax53          0         0          0          0          0          0
#> tax54          0         0          0          0          0          0
#> tax55          0         0          0          0          0          0
#> tax56          0         0          0          0          0          0
#> tax57          0         0          0          0          0          0
#> tax58          0         0          0          0          0          0
#> tax59          0         0          0          0          0          0
#> tax60          0         0          0          0          0          0
#> tax61          0         0          0          0          0          0
#> tax62          0         0          0          0          0          0
#> tax63          0         0          0          0          0          0
#> tax64          0         0          0          0          0          0
#> tax65          0         0          0          0          0          0
#> tax66          0         0          0          0          0          0
#> tax67          0         0          0          0          0          0
#> tax68          0         0          0          0          0          0
#> tax69          0         0          0          0          0          0
#> tax70          0         0          0          0          0          0
#> tax71          0         0          0          0          0          0
#> tax72          0         0          0          0          0          0
#> tax73          0         0          0          0          0          0
#> tax74          0         0          0          0          0          0
#> tax75          0         0          0          0          0          0
#> tax76          0         0          0          0          0          0
#> tax77          0         0          0          0          0          0
#> tax78          0         0          0          0          0          0
#> tax79          0         0          0          0          0          0
#> tax80          0         0          0          0          0          0
#> tax81          0         0          0          0          0          0
#> tax82          0         0          0          0          0          0
#> tax83          0         0          0          0          0          0
#> tax84          0         0          0          0          0          0
#> tax85          0         0          0          0          0          0
#> tax86          0         0          0          0          0          0
#> tax87          0         0          0          0          0          0
#> tax88          0         0          0          0          0          0
#> tax89          0         0          0          0          0          0
#> tax90          0         0          0          0          0          0
#> tax91          0         0          0          0          0          0
#> tax92          0         0          0          0          0          0
#> tax93          0         0          0          0          0          0
#> tax94          0         0          0          0          0          0
#> tax95          0         0          0          0          0          0
#> tax96          0         0          0          0          0          0
#> tax97          0         0          0          0          0          0
#> tax98          0         0          0          0          0          0
#> tax99          0         0          0          0          0          0
#> tax100         0         0          0          0          0          0
#>        S39_S1_T14 S39_S1_T15 S39_S1_T16 S39_S1_T17 S39_S1_T18 S39_S1_T19
#> tax1            0          0          0          0          0          0
#> tax2            0          0          0          0          0          0
#> tax3            0          0          0          0          0          0
#> tax4            0          0          0          0          0          0
#> tax5            0          0          0          0          0          0
#> tax6            0          0          0          0          0          0
#> tax7            0          0          0          0          0          0
#> tax8            0          0          0          0          0          0
#> tax9            0          0          0          0          0          0
#> tax10           0          0          0          0          0          0
#> tax11           0          0          0          0          0          0
#> tax12           0          0          0          0          0          0
#> tax13           0          0          0          0          0          0
#> tax14           0          0          0          0          0          0
#> tax15           0          0          0          0          0          0
#> tax16           0          0          0          0          0          0
#> tax17           0          0          0          0          0          0
#> tax18           0          0          0          0          0          0
#> tax19           0          0          0          0          0          0
#> tax20           0          0          0          0          0          0
#> tax21           0          0          0          0          0          0
#> tax22           0          0          0          0          0          0
#> tax23           0          0          0          0          0          0
#> tax24           0          0          0          0          0          0
#> tax25           0          0          0          0          0          0
#> tax26           0          0          0          0          0          0
#> tax27           0          0          0          0          0          0
#> tax28           0          0          0          0          0          0
#> tax29           0          0          0          0          0          0
#> tax30           0          0          0          0          0          0
#> tax31           0          0          0          0          0          0
#> tax32           0          0          0          0          0          0
#> tax33           0          0          0          0          0          0
#> tax34           0          0          0          0          0          0
#> tax35           0          0          0          0          0          0
#> tax36           0          0          0          0          0          0
#> tax37           0          0          0          0          0          0
#> tax38           0          0          0          0          0          0
#> tax39           0          0          0          0          0          0
#> tax40           0          0          0          0          0          0
#> tax41           0          0          0          0          0          0
#> tax42           0          0          0          0          0          0
#> tax43           0          0          0          0          0          0
#> tax44           0          0          0          0          0          0
#> tax45           0          0          0          0          0          0
#> tax46           0          0          0          0          0          0
#> tax47           0          0          0          0          0          0
#> tax48           0          0          0          0          0          0
#> tax49           0          0          0          0          0          0
#> tax50           0          0          0          0          0          0
#> tax51           0          0          0          0          0          0
#> tax52           0          0          0          0          0          0
#> tax53           0          0          0          0          0          0
#> tax54           0          0          0          0          0          0
#> tax55           0          0          0          0          0          0
#> tax56           0          0          0          0          0          0
#> tax57           0          0          0          0          0          0
#> tax58           0          0          0          0          0          0
#> tax59           0          0          0          0          0          0
#> tax60           0          0          0          0          0          0
#> tax61           0          0          0          0          0          0
#> tax62           0          0          0          0          0          0
#> tax63           0          0          0          0          0          0
#> tax64           0          0          0          0          0          0
#> tax65           0          0          0          0          0          0
#> tax66           0          0          0          0          0          0
#> tax67           0          0          0          0          0          0
#> tax68           0          0          0          0          0          0
#> tax69           0          0          0          0          0          0
#> tax70           0          0          0          0          0          0
#> tax71           0          0          0          0          0          0
#> tax72           0          0          0          0          0          0
#> tax73           0          0          0          0          0          0
#> tax74           0          0          0          0          0          0
#> tax75           0          0          0          0          0          0
#> tax76           0          0          0          0          0          0
#> tax77           0          0          0          0          0          0
#> tax78           0          0          0          0          0          0
#> tax79           0          0          0          0          0          0
#> tax80           0          0          0          0          0          0
#> tax81           0          0          0          0          0          0
#> tax82           0          0          0          0          0          0
#> tax83           0          0          0          0          0          0
#> tax84           0          0          0          0          0          0
#> tax85           0          0          0          0          0          0
#> tax86           0          0          0          0          0          0
#> tax87           0          0          0          0          0          0
#> tax88           0          0          0          0          0          0
#> tax89           0          0          0          0          0          0
#> tax90           0          0          0          0          0          0
#> tax91           0          0          0          0          0          0
#> tax92           0          0          0          0          0          0
#> tax93           0          0          0          0          0          0
#> tax94           0          0          0          0          0          0
#> tax95           0          0          0          0          0          0
#> tax96           0          0          0          0          0          0
#> tax97           0          0          0          0          0          0
#> tax98           0          0          0          0          0          0
#> tax99           0          0          0          0          0          0
#> tax100          0          0          0          0          0          0
#>        S39_S1_T20 S39_S1_T21 S39_S1_T22 S39_S1_T23 S39_S1_T24 S39_S1_T25
#> tax1            0          0          0          0          0          0
#> tax2            0          0          0          0          0          0
#> tax3            0          0          0          0          0          0
#> tax4            0          0          0          0          0          0
#> tax5            0          0          0          0          0          0
#> tax6            0          0          0          0          0          0
#> tax7            0          0          0          0          0          0
#> tax8            0          0          0          0          0          0
#> tax9            0          0          0          0          0          0
#> tax10           0          0          0          0          0          0
#> tax11           0          0          0          0          0          0
#> tax12           0          0          0          0          0          0
#> tax13           0          0          0          0          0          0
#> tax14           0          0          0          0          0          0
#> tax15           0          0          0          0          0          0
#> tax16           0          0          0          0          0          0
#> tax17           0          0          0          0          0          0
#> tax18           0          0          0          0          0          0
#> tax19           0          0          0          0          0          0
#> tax20           0          0          0          0          0          0
#> tax21           0          0          0          0          0          0
#> tax22           0          0          0          0          0          0
#> tax23           0          0          0          0          0          0
#> tax24           0          0          0          0          0          0
#> tax25           0          0          0          0          0          0
#> tax26           0          0          0          0          0          0
#> tax27           0          0          0          0          0          0
#> tax28           0          0          0          0          0          0
#> tax29           0          0          0          0          0          0
#> tax30           0          0          0          0          0          0
#> tax31           0          0          0          0          0          0
#> tax32           0          0          0          0          0          0
#> tax33           0          0          0          0          0          0
#> tax34           0          0          0          0          0          0
#> tax35           0          0          0          0          0          0
#> tax36           0          0          0          0          0          0
#> tax37           0          0          0          0          0          0
#> tax38           0          0          0          0          0          0
#> tax39           0          0          0          0          0          0
#> tax40           0          0          0          0          0          0
#> tax41           0          0          0          0          0          0
#> tax42           0          0          0          0          0          0
#> tax43           0          0          0          0          0          0
#> tax44           0          0          0          0          0          0
#> tax45           0          0          0          0          0          0
#> tax46           0          0          0          0          0          0
#> tax47           0          0          0          0          0          0
#> tax48           0          0          0          0          0          0
#> tax49           0          0          0          0          0          0
#> tax50           0          0          0          0          0          0
#> tax51           0          0          0          0          0          0
#> tax52           0          0          0          0          0          0
#> tax53           0          0          0          0          0          0
#> tax54           0          0          0          0          0          0
#> tax55           0          0          0          0          0          0
#> tax56           0          0          0          0          0          0
#> tax57           0          0          0          0          0          0
#> tax58           0          0          0          0          0          0
#> tax59           0          0          0          0          0          0
#> tax60           0          0          0          0          0          0
#> tax61           0          0          0          0          0          0
#> tax62           0          0          0          0          0          0
#> tax63           0          0          0          0          0          0
#> tax64           0          0          0          0          0          0
#> tax65           0          0          0          0          0          0
#> tax66           0          0          0          0          0          0
#> tax67           0          0          0          0          0          0
#> tax68           0          0          0          0          0          0
#> tax69           0          0          0          0          0          0
#> tax70           0          0          0          0          0          0
#> tax71           0          0          0          0          0          0
#> tax72           0          0          0          0          0          0
#> tax73           0          0          0          0          0          0
#> tax74           0          0          0          0          0          0
#> tax75           0          0          0          0          0          0
#> tax76           0          0          0          0          0          0
#> tax77           0          0          0          0          0          0
#> tax78           0          0          0          0          0          0
#> tax79           0          0          0          0          0          0
#> tax80           0          0          0          0          0          0
#> tax81           0          0          0          0          0          0
#> tax82           0          0          0          0          0          0
#> tax83           0          0          0          0          0          0
#> tax84           0          0          0          0          0          0
#> tax85           0          0          0          0          0          0
#> tax86           0          0          0          0          0          0
#> tax87           0          0          0          0          0          0
#> tax88           0          0          0          0          0          0
#> tax89           0          0          0          0          0          0
#> tax90           0          0          0          0          0          0
#> tax91           0          0          0          0          0          0
#> tax92           0          0          0          0          0          0
#> tax93           0          0          0          0          0          0
#> tax94           0          0          0          0          0          0
#> tax95           0          0          0          0          0          0
#> tax96           0          0          0          0          0          0
#> tax97           0          0          0          0          0          0
#> tax98           0          0          0          0          0          0
#> tax99           0          0          0          0          0          0
#> tax100          0          0          0          0          0          0
#>        S39_S1_T26 S39_S1_T27          S39_         S39_         S39_
#> tax1            0          0 -0.4995692117 -0.822412979 -0.954145698
#> tax2            0          0 -8.5508308865 -8.689295803 -8.908897479
#> tax3            0          0  0.6598739624  0.592588640  0.446615253
#> tax4            0          0  0.1718330383 -0.327391840 -0.509042047
#> tax5            0          0 -4.4094324112 -5.282067980 -5.482053950
#> tax6            0          0 -0.6565771103 -1.081117255 -1.270604679
#> tax7            0          0 -1.9450407028 -1.963119859 -2.147154751
#> tax8            0          0 -2.9156884466 -3.229302917 -3.416328578
#> tax9            0          0 -2.0043707348 -1.820433866 -1.941316468
#> tax10           0          0 -4.3931485925 -4.745469457 -4.949238266
#> tax11           0          0  0.2168416977  0.107768842 -0.045846292
#> tax12           0          0  0.2806835175  0.163122870 -0.020397958
#> tax13           0          0  0.3395166397  0.214177120  0.087142978
#> tax14           0          0  0.0601911545  0.004712355 -0.106491736
#> tax15           0          0  0.1096220016 -0.071225405 -0.247479552
#> tax16           0          0 -0.0168237686 -0.155899411 -0.302792220
#> tax17           0          0  0.3259725571  0.280710334  0.117709478
#> tax18           0          0  0.3351678848  0.215392238  0.071738618
#> tax19           0          0  0.5575590134  0.395278386  0.239038286
#> tax20           0          0  0.7593412399  0.678184589  0.517555816
#> tax21           0          0  0.2238121033  0.085793177 -0.069928930
#> tax22           0          0 -0.0058698654 -0.303897267 -0.479201419
#> tax23           0          0  0.3086833954  0.236229556  0.101684037
#> tax24           0          0  0.1452112198  0.074289276 -0.076928082
#> tax25           0          0  0.3844842911  0.297793036  0.188796520
#> tax26           0          0  0.0006732941 -0.272484813 -0.433027551
#> tax27           0          0  0.0512857437 -0.056699128 -0.203551962
#> tax28           0          0  0.2152891159 -0.106400195 -0.280809130
#> tax29           0          0  0.5388193130  0.523353599  0.410513571
#> tax30           0          0  0.0441179276 -0.033385754 -0.156673295
#> tax31           0          0  0.3237743378  0.184244848  0.067805676
#> tax32           0          0  0.2361926635  0.197840742  0.082751388
#> tax33           0          0  0.2919178009  0.333822125  0.220355636
#> tax34           0          0  0.1786785126  0.071316208 -0.039956337
#> tax35           0          0  0.1506662369  0.020798672 -0.114558390
#> tax36           0          0  0.0180006027 -0.115111294 -0.256816001
#> tax37           0          0  0.1945858002  0.068649633 -0.078086308
#> tax38           0          0  0.0233964920 -0.044654358 -0.170500903
#> tax39           0          0  0.1983098984 -0.015081247 -0.161097992
#> tax40           0          0  0.3344850938  0.266528533  0.158080828
#> tax41           0          0  0.1287546158  0.042823950 -0.087038301
#> tax42           0          0  0.5104541779  0.374932210  0.199022713
#> tax43           0          0 -0.0331935883 -0.248908758 -0.402331171
#> tax44           0          0  0.2297506332  0.046746754 -0.103512957
#> tax45           0          0  0.3941631317  0.270492633  0.160490161
#> tax46           0          0  0.3290901184  0.093485503 -0.062861000
#> tax47           0          0  0.2086553574  0.056777012 -0.096024956
#> tax48           0          0  0.0120363235 -0.090106987 -0.205240159
#> tax49           0          0  0.1354703903  0.139493182  0.009939909
#> tax50           0          0  0.7509851456  0.634220101  0.487003326
#> tax51           0          0 -0.0045952797 -0.097358545 -0.210481723
#> tax52           0          0  0.3034396853  0.166627509  0.026412850
#> tax53           0          0  0.0473241806 -0.009547245 -0.165518170
#> tax54           0          0 -0.1613893509 -0.386489732 -0.521509091
#> tax55           0          0  0.2053546906  0.077565908 -0.086565302
#> tax56           0          0  0.3104982376  0.193796737  0.049180485
#> tax57           0          0  0.4718093872  0.450937078  0.342278571
#> tax58           0          0  0.1600780487  0.022719247 -0.122299035
#> tax59           0          0  0.0146732330 -0.104139680 -0.225836447
#> tax60           0          0  0.2541885376  0.145515919  0.008349135
#> tax61           0          0  0.2945899963  0.375087999  0.196870429
#> tax62           0          0  0.1574106216  0.113596950 -0.008143766
#> tax63           0          0  0.0593261548 -0.053802626 -0.180621732
#> tax64           0          0  0.1407775879 -0.105427583 -0.293193374
#> tax65           0          0  0.4615154266  0.154645556 -0.033891769
#> tax66           0          0  0.3066682816  0.192604213  0.060432207
#> tax67           0          0 -0.2125597000 -0.378993023 -0.526691073
#> tax68           0          0  0.0305347443 -0.150043715 -0.281304019
#> tax69           0          0  0.3402118683  0.282630659  0.128701732
#> tax70           0          0  0.1442337036  0.071835813 -0.051035370
#> tax71           0          0  0.0624885559 -0.057159878 -0.181228229
#> tax72           0          0 -0.2073659897 -0.513534955 -0.672618117
#> tax73           0          0  0.4265193939  0.237494582  0.080594653
#> tax74           0          0  0.2029018402 -0.083552179 -0.237920988
#> tax75           0          0  0.4356718063  0.330175468  0.179106429
#> tax76           0          0  0.2899551392  0.099072195 -0.063719216
#> tax77           0          0  0.0260477066 -0.147750048 -0.276533240
#> tax78           0          0  0.2460842133  0.134428251 -0.001043183
#> tax79           0          0 -0.1348543167 -0.276021333 -0.417208342
#> tax80           0          0  0.0940208435 -0.111041739 -0.296382938
#> tax81           0          0  0.1685562134  0.100798891 -0.001135293
#> tax82           0          0  0.1139745826 -0.058045467 -0.194586493
#> tax83           0          0 -0.0741682053 -0.195786544 -0.368228980
#> tax84           0          0  0.2384967804  0.108518067 -0.047965663
#> tax85           0          0 -0.1305198499 -0.245109575 -0.352930796
#> tax86           0          0  0.3541364670  0.134992781 -0.004320883
#> tax87           0          0  0.3465108871  0.177372683  0.042028859
#> tax88           0          0  0.4910478592  0.359768334  0.145573707
#> tax89           0          0  0.2028779870  0.068472272 -0.059382109
#> tax90           0          0  0.1261796951 -0.002716995 -0.134638230
#> tax91           0          0 -0.1283769608 -0.344594899 -0.487703891
#> tax92           0          0  0.3296861649  0.285835312  0.134617647
#> tax93           0          0  0.1123609543 -0.069674265 -0.210840804
#> tax94           0          0  0.3222050213  0.347940837  0.238310587
#> tax95           0          0 -0.0021104813 -0.119446459 -0.298736413
#> tax96           0          0  0.0517568588 -0.020600251 -0.155873673
#> tax97           0          0  0.4143972283  0.459154946  0.275084348
#> tax98           0          0  0.2685546875  0.127207983 -0.036531074
#> tax99           0          0  0.4917688370  0.259683178  0.089818898
#> tax100          0          0  0.3029069901  0.184510015  0.047597635