Counterfactual Partial Dependence Effects

Our selection algorithm, select_taxa depends on estimates for counterfactual partial dependence effects across all taxa and time lags. This supports estimation of these effects for a single random data split pair.

Usage

pd_effects(
  fit,
  ts,
  w0,
  w1,
  n_sample = NULL,
  patch_len = 8,
  intervention_len = NULL
)

Arguments

fit

An object of class mbtransfer_model, as generated using the mbtransfer function. This includes trained boosting models for every taxon, stored within the @parameters slot.

ts

An object of class ts_inter containing the time-varying microbiome community, environmental interventions, and static host features data. The columns for each element of the values matrix are expected to be consecutive timepoints in an changing community. ts_from_dfs() is a helper to create these objects from data.frames with the relevant information.

w0

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

w1

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

Examples

data(sim_ts)
w0 <- cbind(sim_ts[[1]]@interventions, matrix(0, nrow = 1, ncol = 3))
w1 <- cbind(sim_ts[[1]]@interventions, matrix(1, nrow = 1, ncol = 3))
fit <- mbtransfer(sim_ts, 1, 1, nrounds = 10)
pd_effects(fit, sim_ts, w0, w1)
#>        S8_S1_T1 S8_S1_T2 S8_S1_T3 S8_S1_T4 S8_S1_T5 S8_S1_T6 S8_S1_T7 S8_S1_T8
#> tax1          0        0        0        0        0        0        0        0
#> tax2          0        0        0        0        0        0        0        0
#> tax3          0        0        0        0        0        0        0        0
#> tax4          0        0        0        0        0        0        0        0
#> tax5          0        0        0        0        0        0        0        0
#> tax6          0        0        0        0        0        0        0        0
#> tax7          0        0        0        0        0        0        0        0
#> tax8          0        0        0        0        0        0        0        0
#> tax9          0        0        0        0        0        0        0        0
#> tax10         0        0        0        0        0        0        0        0
#> tax11         0        0        0        0        0        0        0        0
#> tax12         0        0        0        0        0        0        0        0
#> tax13         0        0        0        0        0        0        0        0
#> tax14         0        0        0        0        0        0        0        0
#> tax15         0        0        0        0        0        0        0        0
#> tax16         0        0        0        0        0        0        0        0
#> tax17         0        0        0        0        0        0        0        0
#> tax18         0        0        0        0        0        0        0        0
#> tax19         0        0        0        0        0        0        0        0
#> tax20         0        0        0        0        0        0        0        0
#> tax21         0        0        0        0        0        0        0        0
#> tax22         0        0        0        0        0        0        0        0
#> tax23         0        0        0        0        0        0        0        0
#> tax24         0        0        0        0        0        0        0        0
#> tax25         0        0        0        0        0        0        0        0
#> tax26         0        0        0        0        0        0        0        0
#> tax27         0        0        0        0        0        0        0        0
#> tax28         0        0        0        0        0        0        0        0
#> tax29         0        0        0        0        0        0        0        0
#> tax30         0        0        0        0        0        0        0        0
#> tax31         0        0        0        0        0        0        0        0
#> tax32         0        0        0        0        0        0        0        0
#> tax33         0        0        0        0        0        0        0        0
#> tax34         0        0        0        0        0        0        0        0
#> tax35         0        0        0        0        0        0        0        0
#> tax36         0        0        0        0        0        0        0        0
#> tax37         0        0        0        0        0        0        0        0
#> tax38         0        0        0        0        0        0        0        0
#> tax39         0        0        0        0        0        0        0        0
#> tax40         0        0        0        0        0        0        0        0
#> tax41         0        0        0        0        0        0        0        0
#> tax42         0        0        0        0        0        0        0        0
#> tax43         0        0        0        0        0        0        0        0
#> tax44         0        0        0        0        0        0        0        0
#> tax45         0        0        0        0        0        0        0        0
#> tax46         0        0        0        0        0        0        0        0
#> tax47         0        0        0        0        0        0        0        0
#> tax48         0        0        0        0        0        0        0        0
#> tax49         0        0        0        0        0        0        0        0
#> tax50         0        0        0        0        0        0        0        0
#> tax51         0        0        0        0        0        0        0        0
#> tax52         0        0        0        0        0        0        0        0
#> tax53         0        0        0        0        0        0        0        0
#> tax54         0        0        0        0        0        0        0        0
#> tax55         0        0        0        0        0        0        0        0
#> tax56         0        0        0        0        0        0        0        0
#> tax57         0        0        0        0        0        0        0        0
#> tax58         0        0        0        0        0        0        0        0
#> tax59         0        0        0        0        0        0        0        0
#> tax60         0        0        0        0        0        0        0        0
#> tax61         0        0        0        0        0        0        0        0
#> tax62         0        0        0        0        0        0        0        0
#> tax63         0        0        0        0        0        0        0        0
#> tax64         0        0        0        0        0        0        0        0
#> tax65         0        0        0        0        0        0        0        0
#> tax66         0        0        0        0        0        0        0        0
#> tax67         0        0        0        0        0        0        0        0
#> tax68         0        0        0        0        0        0        0        0
#> tax69         0        0        0        0        0        0        0        0
#> tax70         0        0        0        0        0        0        0        0
#> tax71         0        0        0        0        0        0        0        0
#> tax72         0        0        0        0        0        0        0        0
#> tax73         0        0        0        0        0        0        0        0
#> tax74         0        0        0        0        0        0        0        0
#> tax75         0        0        0        0        0        0        0        0
#> tax76         0        0        0        0        0        0        0        0
#> tax77         0        0        0        0        0        0        0        0
#> tax78         0        0        0        0        0        0        0        0
#> tax79         0        0        0        0        0        0        0        0
#> tax80         0        0        0        0        0        0        0        0
#> tax81         0        0        0        0        0        0        0        0
#> tax82         0        0        0        0        0        0        0        0
#> tax83         0        0        0        0        0        0        0        0
#> tax84         0        0        0        0        0        0        0        0
#> tax85         0        0        0        0        0        0        0        0
#> tax86         0        0        0        0        0        0        0        0
#> tax87         0        0        0        0        0        0        0        0
#> tax88         0        0        0        0        0        0        0        0
#> tax89         0        0        0        0        0        0        0        0
#> tax90         0        0        0        0        0        0        0        0
#> tax91         0        0        0        0        0        0        0        0
#> tax92         0        0        0        0        0        0        0        0
#> tax93         0        0        0        0        0        0        0        0
#> tax94         0        0        0        0        0        0        0        0
#> tax95         0        0        0        0        0        0        0        0
#> tax96         0        0        0        0        0        0        0        0
#> tax97         0        0        0        0        0        0        0        0
#> tax98         0        0        0        0        0        0        0        0
#> tax99         0        0        0        0        0        0        0        0
#> tax100        0        0        0        0        0        0        0        0
#>        S8_S1_T9 S8_S1_T10 S8_S1_T11 S8_S1_T12 S8_S1_T13 S8_S1_T14 S8_S1_T15
#> tax1          0         0         0         0         0         0         0
#> tax2          0         0         0         0         0         0         0
#> tax3          0         0         0         0         0         0         0
#> tax4          0         0         0         0         0         0         0
#> tax5          0         0         0         0         0         0         0
#> tax6          0         0         0         0         0         0         0
#> tax7          0         0         0         0         0         0         0
#> tax8          0         0         0         0         0         0         0
#> tax9          0         0         0         0         0         0         0
#> tax10         0         0         0         0         0         0         0
#> tax11         0         0         0         0         0         0         0
#> tax12         0         0         0         0         0         0         0
#> tax13         0         0         0         0         0         0         0
#> tax14         0         0         0         0         0         0         0
#> tax15         0         0         0         0         0         0         0
#> tax16         0         0         0         0         0         0         0
#> tax17         0         0         0         0         0         0         0
#> tax18         0         0         0         0         0         0         0
#> tax19         0         0         0         0         0         0         0
#> tax20         0         0         0         0         0         0         0
#> tax21         0         0         0         0         0         0         0
#> tax22         0         0         0         0         0         0         0
#> tax23         0         0         0         0         0         0         0
#> tax24         0         0         0         0         0         0         0
#> tax25         0         0         0         0         0         0         0
#> tax26         0         0         0         0         0         0         0
#> tax27         0         0         0         0         0         0         0
#> tax28         0         0         0         0         0         0         0
#> tax29         0         0         0         0         0         0         0
#> tax30         0         0         0         0         0         0         0
#> tax31         0         0         0         0         0         0         0
#> tax32         0         0         0         0         0         0         0
#> tax33         0         0         0         0         0         0         0
#> tax34         0         0         0         0         0         0         0
#> tax35         0         0         0         0         0         0         0
#> tax36         0         0         0         0         0         0         0
#> tax37         0         0         0         0         0         0         0
#> tax38         0         0         0         0         0         0         0
#> tax39         0         0         0         0         0         0         0
#> tax40         0         0         0         0         0         0         0
#> tax41         0         0         0         0         0         0         0
#> tax42         0         0         0         0         0         0         0
#> tax43         0         0         0         0         0         0         0
#> tax44         0         0         0         0         0         0         0
#> tax45         0         0         0         0         0         0         0
#> tax46         0         0         0         0         0         0         0
#> tax47         0         0         0         0         0         0         0
#> tax48         0         0         0         0         0         0         0
#> tax49         0         0         0         0         0         0         0
#> tax50         0         0         0         0         0         0         0
#> tax51         0         0         0         0         0         0         0
#> tax52         0         0         0         0         0         0         0
#> tax53         0         0         0         0         0         0         0
#> tax54         0         0         0         0         0         0         0
#> tax55         0         0         0         0         0         0         0
#> tax56         0         0         0         0         0         0         0
#> tax57         0         0         0         0         0         0         0
#> tax58         0         0         0         0         0         0         0
#> tax59         0         0         0         0         0         0         0
#> tax60         0         0         0         0         0         0         0
#> tax61         0         0         0         0         0         0         0
#> tax62         0         0         0         0         0         0         0
#> tax63         0         0         0         0         0         0         0
#> tax64         0         0         0         0         0         0         0
#> tax65         0         0         0         0         0         0         0
#> tax66         0         0         0         0         0         0         0
#> tax67         0         0         0         0         0         0         0
#> tax68         0         0         0         0         0         0         0
#> tax69         0         0         0         0         0         0         0
#> tax70         0         0         0         0         0         0         0
#> tax71         0         0         0         0         0         0         0
#> tax72         0         0         0         0         0         0         0
#> tax73         0         0         0         0         0         0         0
#> tax74         0         0         0         0         0         0         0
#> tax75         0         0         0         0         0         0         0
#> tax76         0         0         0         0         0         0         0
#> tax77         0         0         0         0         0         0         0
#> tax78         0         0         0         0         0         0         0
#> tax79         0         0         0         0         0         0         0
#> tax80         0         0         0         0         0         0         0
#> tax81         0         0         0         0         0         0         0
#> tax82         0         0         0         0         0         0         0
#> tax83         0         0         0         0         0         0         0
#> tax84         0         0         0         0         0         0         0
#> tax85         0         0         0         0         0         0         0
#> tax86         0         0         0         0         0         0         0
#> tax87         0         0         0         0         0         0         0
#> tax88         0         0         0         0         0         0         0
#> tax89         0         0         0         0         0         0         0
#> tax90         0         0         0         0         0         0         0
#> tax91         0         0         0         0         0         0         0
#> tax92         0         0         0         0         0         0         0
#> tax93         0         0         0         0         0         0         0
#> tax94         0         0         0         0         0         0         0
#> tax95         0         0         0         0         0         0         0
#> tax96         0         0         0         0         0         0         0
#> tax97         0         0         0         0         0         0         0
#> tax98         0         0         0         0         0         0         0
#> tax99         0         0         0         0         0         0         0
#> tax100        0         0         0         0         0         0         0
#>        S8_S1_T16 S8_S1_T17 S8_S1_T18 S8_S1_T19 S8_S1_T20 S8_S1_T21 S8_S1_T22
#> tax1           0         0         0         0         0         0         0
#> tax2           0         0         0         0         0         0         0
#> tax3           0         0         0         0         0         0         0
#> tax4           0         0         0         0         0         0         0
#> tax5           0         0         0         0         0         0         0
#> tax6           0         0         0         0         0         0         0
#> tax7           0         0         0         0         0         0         0
#> tax8           0         0         0         0         0         0         0
#> tax9           0         0         0         0         0         0         0
#> tax10          0         0         0         0         0         0         0
#> tax11          0         0         0         0         0         0         0
#> tax12          0         0         0         0         0         0         0
#> tax13          0         0         0         0         0         0         0
#> tax14          0         0         0         0         0         0         0
#> tax15          0         0         0         0         0         0         0
#> tax16          0         0         0         0         0         0         0
#> tax17          0         0         0         0         0         0         0
#> tax18          0         0         0         0         0         0         0
#> tax19          0         0         0         0         0         0         0
#> tax20          0         0         0         0         0         0         0
#> tax21          0         0         0         0         0         0         0
#> tax22          0         0         0         0         0         0         0
#> tax23          0         0         0         0         0         0         0
#> tax24          0         0         0         0         0         0         0
#> tax25          0         0         0         0         0         0         0
#> tax26          0         0         0         0         0         0         0
#> tax27          0         0         0         0         0         0         0
#> tax28          0         0         0         0         0         0         0
#> tax29          0         0         0         0         0         0         0
#> tax30          0         0         0         0         0         0         0
#> tax31          0         0         0         0         0         0         0
#> tax32          0         0         0         0         0         0         0
#> tax33          0         0         0         0         0         0         0
#> tax34          0         0         0         0         0         0         0
#> tax35          0         0         0         0         0         0         0
#> tax36          0         0         0         0         0         0         0
#> tax37          0         0         0         0         0         0         0
#> tax38          0         0         0         0         0         0         0
#> tax39          0         0         0         0         0         0         0
#> tax40          0         0         0         0         0         0         0
#> tax41          0         0         0         0         0         0         0
#> tax42          0         0         0         0         0         0         0
#> tax43          0         0         0         0         0         0         0
#> tax44          0         0         0         0         0         0         0
#> tax45          0         0         0         0         0         0         0
#> tax46          0         0         0         0         0         0         0
#> tax47          0         0         0         0         0         0         0
#> tax48          0         0         0         0         0         0         0
#> tax49          0         0         0         0         0         0         0
#> tax50          0         0         0         0         0         0         0
#> tax51          0         0         0         0         0         0         0
#> tax52          0         0         0         0         0         0         0
#> tax53          0         0         0         0         0         0         0
#> tax54          0         0         0         0         0         0         0
#> tax55          0         0         0         0         0         0         0
#> tax56          0         0         0         0         0         0         0
#> tax57          0         0         0         0         0         0         0
#> tax58          0         0         0         0         0         0         0
#> tax59          0         0         0         0         0         0         0
#> tax60          0         0         0         0         0         0         0
#> tax61          0         0         0         0         0         0         0
#> tax62          0         0         0         0         0         0         0
#> tax63          0         0         0         0         0         0         0
#> tax64          0         0         0         0         0         0         0
#> tax65          0         0         0         0         0         0         0
#> tax66          0         0         0         0         0         0         0
#> tax67          0         0         0         0         0         0         0
#> tax68          0         0         0         0         0         0         0
#> tax69          0         0         0         0         0         0         0
#> tax70          0         0         0         0         0         0         0
#> tax71          0         0         0         0         0         0         0
#> tax72          0         0         0         0         0         0         0
#> tax73          0         0         0         0         0         0         0
#> tax74          0         0         0         0         0         0         0
#> tax75          0         0         0         0         0         0         0
#> tax76          0         0         0         0         0         0         0
#> tax77          0         0         0         0         0         0         0
#> tax78          0         0         0         0         0         0         0
#> tax79          0         0         0         0         0         0         0
#> tax80          0         0         0         0         0         0         0
#> tax81          0         0         0         0         0         0         0
#> tax82          0         0         0         0         0         0         0
#> tax83          0         0         0         0         0         0         0
#> tax84          0         0         0         0         0         0         0
#> tax85          0         0         0         0         0         0         0
#> tax86          0         0         0         0         0         0         0
#> tax87          0         0         0         0         0         0         0
#> tax88          0         0         0         0         0         0         0
#> tax89          0         0         0         0         0         0         0
#> tax90          0         0         0         0         0         0         0
#> tax91          0         0         0         0         0         0         0
#> tax92          0         0         0         0         0         0         0
#> tax93          0         0         0         0         0         0         0
#> tax94          0         0         0         0         0         0         0
#> tax95          0         0         0         0         0         0         0
#> tax96          0         0         0         0         0         0         0
#> tax97          0         0         0         0         0         0         0
#> tax98          0         0         0         0         0         0         0
#> tax99          0         0         0         0         0         0         0
#> tax100         0         0         0         0         0         0         0
#>        S8_S1_T23 S8_S1_T24 S8_S1_T25 S8_S1_T26 S8_S1_T27           S8_
#> tax1           0         0         0         0         0 -0.6328802222
#> tax2           0         0         0         0         0 -8.8437741257
#> tax3           0         0         0         0         0  0.4410572052
#> tax4           0         0         0         0         0  0.0174821445
#> tax5           0         0         0         0         0 -4.5244598389
#> tax6           0         0         0         0         0 -0.8993101120
#> tax7           0         0         0         0         0 -2.2115803900
#> tax8           0         0         0         0         0 -3.1882864407
#> tax9           0         0         0         0         0 -2.2448129654
#> tax10          0         0         0         0         0 -4.6580564295
#> tax11          0         0         0         0         0  0.0047330856
#> tax12          0         0         0         0         0  0.0189847946
#> tax13          0         0         0         0         0  0.1657667160
#> tax14          0         0         0         0         0 -0.0562362330
#> tax15          0         0         0         0         0 -0.0525388718
#> tax16          0         0         0         0         0 -0.2123460770
#> tax17          0         0         0         0         0  0.0883607864
#> tax18          0         0         0         0         0  0.1418313980
#> tax19          0         0         0         0         0  0.3188953400
#> tax20          0         0         0         0         0  0.5106287003
#> tax21          0         0         0         0         0  0.0186796188
#> tax22          0         0         0         0         0 -0.1705785819
#> tax23          0         0         0         0         0  0.1213951111
#> tax24          0         0         0         0         0 -0.0383281708
#> tax25          0         0         0         0         0  0.2016544342
#> tax26          0         0         0         0         0 -0.1238431931
#> tax27          0         0         0         0         0 -0.1185169220
#> tax28          0         0         0         0         0  0.0637636185
#> tax29          0         0         0         0         0  0.3474597931
#> tax30          0         0         0         0         0 -0.0841779709
#> tax31          0         0         0         0         0  0.1461267471
#> tax32          0         0         0         0         0  0.1200756345
#> tax33          0         0         0         0         0  0.1051912308
#> tax34          0         0         0         0         0  0.0503978729
#> tax35          0         0         0         0         0  0.0007390976
#> tax36          0         0         0         0         0 -0.1462259293
#> tax37          0         0         0         0         0  0.0016946793
#> tax38          0         0         0         0         0 -0.0815124512
#> tax39          0         0         0         0         0 -0.0050592422
#> tax40          0         0         0         0         0  0.1814304988
#> tax41          0         0         0         0         0  0.0000000000
#> tax42          0         0         0         0         0  0.2495574951
#> tax43          0         0         0         0         0 -0.2750339508
#> tax44          0         0         0         0         0  0.0548744202
#> tax45          0         0         0         0         0  0.2248420488
#> tax46          0         0         0         0         0  0.1524724960
#> tax47          0         0         0         0         0  0.0408697128
#> tax48          0         0         0         0         0 -0.1416769028
#> tax49          0         0         0         0         0 -0.0486803055
#> tax50          0         0         0         0         0  0.5374212265
#> tax51          0         0         0         0         0 -0.1485261917
#> tax52          0         0         0         0         0  0.1151742935
#> tax53          0         0         0         0         0 -0.1205968857
#> tax54          0         0         0         0         0 -0.3675680161
#> tax55          0         0         0         0         0  0.0257902145
#> tax56          0         0         0         0         0  0.1291131973
#> tax57          0         0         0         0         0  0.2687921524
#> tax58          0         0         0         0         0 -0.0005226135
#> tax59          0         0         0         0         0 -0.1095457077
#> tax60          0         0         0         0         0  0.0369777679
#> tax61          0         0         0         0         0 -0.0030059814
#> tax62          0         0         0         0         0  0.0065469742
#> tax63          0         0         0         0         0 -0.0619367645
#> tax64          0         0         0         0         0 -0.0077524185
#> tax65          0         0         0         0         0  0.2494735718
#> tax66          0         0         0         0         0  0.1083784103
#> tax67          0         0         0         0         0 -0.4390964508
#> tax68          0         0         0         0         0 -0.0703191757
#> tax69          0         0         0         0         0  0.1078443527
#> tax70          0         0         0         0         0  0.0015411377
#> tax71          0         0         0         0         0 -0.0470839796
#> tax72          0         0         0         0         0 -0.4460372925
#> tax73          0         0         0         0         0  0.2072572708
#> tax74          0         0         0         0         0  0.0500469208
#> tax75          0         0         0         0         0  0.1642303467
#> tax76          0         0         0         0         0  0.0569200516
#> tax77          0         0         0         0         0 -0.0892047882
#> tax78          0         0         0         0         0  0.0475702286
#> tax79          0         0         0         0         0 -0.3316793442
#> tax80          0         0         0         0         0 -0.0647096634
#> tax81          0         0         0         0         0  0.0049247742
#> tax82          0         0         0         0         0 -0.0319490433
#> tax83          0         0         0         0         0 -0.2685098648
#> tax84          0         0         0         0         0  0.0190181732
#> tax85          0         0         0         0         0 -0.2951179062
#> tax86          0         0         0         0         0  0.1837358475
#> tax87          0         0         0         0         0  0.1630449295
#> tax88          0         0         0         0         0  0.2348232269
#> tax89          0         0         0         0         0  0.0548496246
#> tax90          0         0         0         0         0 -0.0032720566
#> tax91          0         0         0         0         0 -0.3443593979
#> tax92          0         0         0         0         0  0.1289834976
#> tax93          0         0         0         0         0 -0.0346021652
#> tax94          0         0         0         0         0  0.1041212593
#> tax95          0         0         0         0         0 -0.2353229523
#> tax96          0         0         0         0         0 -0.0891284943
#> tax97          0         0         0         0         0  0.0640258789
#> tax98          0         0         0         0         0  0.0483798981
#> tax99          0         0         0         0         0  0.2897911072
#> tax100         0         0         0         0         0  0.0845527649
#>                 S8_           S8_
#> tax1   -0.841968565 -0.8615373997
#> tax2   -8.616192341 -8.6385455018
#> tax3    0.535736538  0.5505777654
#> tax4   -0.381647235 -0.4142778828
#> tax5   -5.305637314 -5.3460654758
#> tax6   -1.183096080 -1.2109350704
#> tax7   -2.026250692 -2.0064143113
#> tax8   -3.337658576 -3.3457967554
#> tax9   -1.902098962 -1.8893083277
#> tax10  -4.784692367 -4.7990008422
#> tax11   0.031126897  0.0330958593
#> tax12   0.074788650  0.0940598193
#> tax13   0.172989561  0.1804017794
#> tax14  -0.015602543 -0.0097186111
#> tax15  -0.089776754 -0.0972206593
#> tax16  -0.212339742 -0.2150062947
#> tax17   0.197786138  0.1936044125
#> tax18   0.115214484  0.0906007630
#> tax19   0.330123856  0.3510301227
#> tax20   0.534867798  0.5215133485
#> tax21   0.020885490  0.0116162073
#> tax22  -0.369475319 -0.3869218486
#> tax23   0.124769744  0.1390411400
#> tax24   0.063179811  0.0520120916
#> tax25   0.227075895  0.2494729360
#> tax26  -0.262710480 -0.2874254045
#> tax27  -0.091604766 -0.0871378013
#> tax28  -0.156010378 -0.1709841319
#> tax29   0.482207423  0.5002813453
#> tax30  -0.032174065 -0.0339026565
#> tax31   0.069299005  0.0956573486
#> tax32   0.132169871  0.1327987058
#> tax33   0.268895694  0.2745615982
#> tax34   0.037475092  0.0316117661
#> tax35  -0.005464111 -0.0144090652
#> tax36  -0.123176518 -0.1220261142
#> tax37   0.019926730  0.0482657523
#> tax38  -0.029512689 -0.0230897041
#> tax39  -0.050295773 -0.0361498538
#> tax40   0.223427387  0.2303639026
#> tax41   0.030535494  0.0170458612
#> tax42   0.286170335  0.2764419942
#> tax43  -0.346275330 -0.3348737898
#> tax44  -0.021795716 -0.0378965991
#> tax45   0.188128778  0.2004478432
#> tax46   0.044671808  0.0412579832
#> tax47   0.028115341  0.0123067470
#> tax48  -0.115365517 -0.1090939158
#> tax49   0.108508916  0.1215218249
#> tax50   0.569268931  0.5544033164
#> tax51  -0.139864939 -0.1608895347
#> tax52   0.108528364  0.1145164172
#> tax53  -0.026961724 -0.0231522946
#> tax54  -0.447786263 -0.4437057745
#> tax55   0.033789828  0.0060634613
#> tax56   0.156848170  0.1539199579
#> tax57   0.401826677  0.4305925823
#> tax58   0.013799054  0.0073995931
#> tax59  -0.109046970 -0.1095652807
#> tax60   0.095507667  0.1076421738
#> tax61   0.202275946  0.2495052814
#> tax62   0.028492825  0.0476537091
#> tax63  -0.098159552 -0.1053551038
#> tax64  -0.118370737 -0.1503392174
#> tax65   0.063651164  0.0428808530
#> tax66   0.089641662  0.0824896608
#> tax67  -0.456824121 -0.4628195990
#> tax68  -0.160371860 -0.1733627433
#> tax69   0.197886444  0.2083867277
#> tax70   0.009934085  0.0117039908
#> tax71  -0.077762059 -0.0819431373
#> tax72  -0.575128374 -0.5797934305
#> tax73   0.174955868  0.1707061813
#> tax74  -0.127778598 -0.1402542364
#> tax75   0.213977155  0.2098576682
#> tax76   0.013142052 -0.0001757145
#> tax77  -0.135754392 -0.1423047270
#> tax78   0.040299382  0.0505805470
#> tax79  -0.331365143 -0.3401576337
#> tax80  -0.123183057 -0.1555313950
#> tax81   0.055552324  0.0761967841
#> tax82  -0.062530132 -0.0443527926
#> tax83  -0.274571260 -0.2915753410
#> tax84   0.006102903 -0.0024762948
#> tax85  -0.311122145 -0.3090155976
#> tax86   0.096155189  0.0922250294
#> tax87   0.133097353  0.1267574969
#> tax88   0.273110594  0.2629119782
#> tax89   0.018138556  0.0089195115
#> tax90  -0.023081076 -0.0269181274
#> tax91  -0.396389257 -0.3952219486
#> tax92   0.255244800  0.2551714806
#> tax93  -0.097639220 -0.0872822489
#> tax94   0.241217755  0.2475347178
#> tax95  -0.178711517 -0.1803448995
#> tax96  -0.036196606 -0.0305729821
#> tax97   0.359314657  0.3695857979
#> tax98   0.075919185  0.0639193058
#> tax99   0.182476418  0.1737608001
#> tax100  0.145250786  0.1461371921