Mirror Splits from Partial Dependence

For ech split, we return an array of dimension n_taxa x 2 fits x n_lags. Each entry contains the estimated counterfactual effect for that taxon and lag combination across the two models (each fit on a different random sample of data).

Usage

pd_splits(ts, w0, w1, tr_fun, n_splits = 20, ...)

Arguments

ts

An object of class ts_inter containing the time-varying microbiome community, environmental interventions, and static host features data. The columns for each element of the values matrix are expected to be consecutive timepoints in an changing community. ts_from_dfs() is a helper to create these objects from data.frames with the relevant information.

w0

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

w1

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

tr_fun

A function that can be used to train over random splits. In the examples in this package, we use an anonymous function that fills our chosen hyperparameters in mbtransfer. For example \(x) mbtransfer(x, P = 1, Q = 1) will fit a lag-1 transfer function model on all the random splits.

Examples

data(sim_ts)
w0 <- cbind(sim_ts[[1]]@interventions, matrix(0, nrow = 1, ncol = 3))
w1 <- cbind(sim_ts[[1]]@interventions, matrix(1, nrow = 1, ncol = 3))
pd_splits(sim_ts, w0, w1, \(x) mbtransfer(x, 1, 1, nrounds = 10), n_splits = 1)
#> Training models for split 1/1
#> [[1]]
#> , , 1
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 2
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 3
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 4
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 5
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 6
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 7
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 8
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 9
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 10
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 11
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 12
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 13
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 14
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 15
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 16
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 17
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 18
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 19
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 20
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 21
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 22
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 23
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 24
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 25
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 26
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 27
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 28
#> 
#>                [,1]         [,2]
#>  [1,]  -2.503260612 -2.962018967
#>  [2,]   3.264472223  3.297410727
#>  [3,]  -0.735269547 -2.127716609
#>  [4,] -10.446120784 -9.059655394
#>  [5,]   3.024178482  3.371935844
#>  [6,]   1.011838436  1.198848724
#>  [7,]  -1.458156586 -1.990365074
#>  [8,]   0.130998611  0.635434151
#>  [9,]   0.120156288 -0.132050469
#> [10,]  -0.008447613 -0.393866539
#> [11,]   0.239701759  0.140673172
#> [12,]  -0.090509415 -0.070543289
#> [13,]  -0.008911133 -0.213920593
#> [14,]  -0.295273781 -0.291205872
#> [15,]  -0.019685745 -0.195842743
#> [16,]   0.041535378  0.274580535
#> [17,]  -0.747425079  0.014483452
#> [18,]   0.120258331 -0.614693142
#> [19,]  -0.230549812 -0.449137688
#> [20,]   0.107275009  0.274610508
#> [21,]  -0.196693420  0.211721420
#> [22,]  -0.657207489 -1.204439731
#> [23,]   0.476626396  0.318047455
#> [24,]   0.250779152 -0.006083534
#> [25,]   0.246780396  0.013621330
#> [26,]  -0.044532776 -0.012242340
#> [27,]   0.388084173 -0.110664368
#> [28,]  -0.226234436 -0.007517815
#> [29,]  -0.006324768 -0.044220470
#> [30,]  -0.392305374 -0.709945679
#> [31,]  -0.795208931 -0.492774010
#> [32,]  -0.056056023 -0.083766937
#> [33,]   0.394457794  0.189302444
#> [34,]   0.365854297  0.185000442
#> [35,]  -0.218740418 -0.300794601
#> [36,]  -1.652155876 -1.396125793
#> [37,]  -0.041197754 -0.194473267
#> [38,]  -1.015259743 -0.141261578
#> [39,]  -0.003720284  0.023448944
#> [40,]  -0.589611053  0.029392242
#> [41,]  -0.002782822 -0.337804794
#> [42,]  -1.194681168 -0.807587692
#> [43,]  -0.191946030  0.333600044
#> [44,]  -0.309320450  0.147950172
#> [45,]  -0.426143646 -0.617706299
#> [46,]  -0.246939636  0.610196091
#> [47,]   0.317388535  0.643639587
#> [48,]  -0.319853783 -0.021185852
#> [49,]  -0.079189301 -0.013333321
#> [50,]  -0.349800768 -0.875688825
#> 
#> , , 29
#> 
#>                [,1]         [,2]
#>  [1,] -3.312112e+00  -4.08782346
#>  [2,]  3.630710e+00   4.08153560
#>  [3,] -1.365837e+00  -2.89933983
#>  [4,] -1.136459e+01 -10.16858035
#>  [5,]  3.142694e+00   3.63670112
#>  [6,]  1.004212e+00   1.20304167
#>  [7,] -2.041837e+00  -2.59418540
#>  [8,]  3.384613e-02   0.64526837
#>  [9,]  4.202255e-02  -0.11804177
#> [10,] -1.448342e-01  -0.41951640
#> [11,]  1.140709e-01   0.01160076
#> [12,] -2.925389e-01  -0.49690935
#> [13,] -2.469494e-01  -0.36009541
#> [14,] -3.179748e-01  -0.41273799
#> [15,]  2.412596e-02  -0.14237540
#> [16,]  2.324213e-01   0.69229334
#> [17,] -7.550888e-01   0.19010641
#> [18,] -3.687937e-01  -0.72595217
#> [19,] -9.190093e-01  -0.81871360
#> [20,]  5.625766e-02   0.29310548
#> [21,] -1.136563e-01   0.16497126
#> [22,] -7.390101e-01  -1.36700178
#> [23,]  4.347692e-01   0.42991071
#> [24,]  1.566215e-02  -0.28501931
#> [25,] -1.409051e-01  -0.16351402
#> [26,] -1.916167e-01  -0.05684117
#> [27,]  1.903349e-01  -0.16832675
#> [28,] -3.819619e-01  -0.04988559
#> [29,] -1.785126e-01  -0.15025037
#> [30,] -4.727248e-01  -0.72351281
#> [31,] -9.815917e-01  -0.66908332
#> [32,] -2.337280e-02   0.07713032
#> [33,]  2.199883e-01   0.18167539
#> [34,]  1.388925e-01   0.35039117
#> [35,] -2.798119e-01  -0.47240273
#> [36,] -2.135446e+00  -1.80274137
#> [37,] -2.855277e-01  -0.41583704
#> [38,] -1.173464e+00  -0.12525792
#> [39,] -7.180291e-02  -0.07240341
#> [40,] -1.085667e+00  -0.44221006
#> [41,] -3.915287e-04  -0.36241570
#> [42,] -1.695493e+00  -1.57555362
#> [43,] -1.141832e-01   0.43760306
#> [44,] -3.437747e-01   0.04092008
#> [45,] -8.889043e-01  -1.27927564
#> [46,] -1.055758e-01   0.59883031
#> [47,]  1.928733e-01   0.44939214
#> [48,] -1.978076e-01  -0.06694776
#> [49,] -1.052628e-01  -0.03062176
#> [50,] -9.750557e-01  -1.73656713
#> 
#> , , 30
#> 
#>                [,1]         [,2]
#>  [1,]  -3.598022030  -4.39711993
#>  [2,]   3.576658964   4.16635764
#>  [3,]  -1.621023360  -3.17262061
#>  [4,] -11.703380153 -10.53061840
#>  [5,]   2.957130705   3.59458106
#>  [6,]   0.880759296   1.12979436
#>  [7,]  -2.275954655  -2.80050207
#>  [8,]  -0.127094496   0.53834895
#>  [9,]  -0.095921902  -0.19097857
#> [10,]  -0.302447955  -0.53099042
#> [11,]  -0.009048212  -0.10135777
#> [12,]  -0.472739878  -0.70543407
#> [13,]  -0.424829256  -0.46451328
#> [14,]  -0.427480414  -0.50080355
#> [15,]  -0.139298802  -0.21100700
#> [16,]   0.168142693   0.72818438
#> [17,]  -0.870327064   0.12152050
#> [18,]  -0.618634269  -0.85492979
#> [19,]  -1.117674396  -1.00097377
#> [20,]  -0.081779934   0.22023148
#> [21,]  -0.183948108   0.05576111
#> [22,]  -0.876403786  -1.52983134
#> [23,]   0.260806674   0.37300298
#> [24,]  -0.172993705  -0.44288306
#> [25,]  -0.304747127  -0.29338392
#> [26,]  -0.338609423  -0.16127986
#> [27,]   0.072560367  -0.25155279
#> [28,]  -0.514365855  -0.14382446
#> [29,]  -0.261550517  -0.22240213
#> [30,]  -0.608885447  -0.81941289
#> [31,]  -1.141732057  -0.79816936
#> [32,]  -0.135829948   0.02375162
#> [33,]   0.057043348   0.06126981
#> [34,]  -0.001661255   0.29210163
#> [35,]  -0.399343195  -0.57422472
#> [36,]  -2.348440851  -1.98937743
#> [37,]  -0.490563256  -0.53124026
#> [38,]  -1.302809238  -0.23553081
#> [39,]  -0.241530714  -0.20493244
#> [40,]  -1.371214821  -0.67823394
#> [41,]  -0.178984665  -0.48212694
#> [42,]  -1.912575722  -1.79007714
#> [43,]  -0.208600567   0.33544220
#> [44,]  -0.462481907  -0.05735077
#> [45,]  -1.112066269  -1.51082223
#> [46,]  -0.240961234   0.49144475
#> [47,]   0.024461837   0.30176923
#> [48,]  -0.330423015  -0.14649051
#> [49,]  -0.215730962  -0.13984678
#> [50,]  -1.218233347  -1.96020674
#> 
#>