Significant Taxa using Mirror Statistics

This selects taxa through a data splitting procedure. It retrains an mbtransfer model across random splits of the data. Partial dependence profiles for pairs of splits are compared with one another -- if they agree for a given feature, then that feature is considered more likely to have a true effect on the response. This function only supports inference of the effects of interventions on taxa responses, but the same principle could apply to estimate significant relationships between taxa.

Usage

select_taxa(ts, w0, w1, tr_fun, qvalue = 0.2, ...)

Arguments

ts

An object of class ts_inter containing the time-varying microbiome community, environmental interventions, and static host features data. The columns for each element of the values matrix are expected to be consecutive timepoints in an changing community. ts_from_dfs() is a helper to create these objects from data.frames with the relevant information.

w0

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

w1

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

tr_fun

A function that can be used to train over random splits. In the examples in this package, we use an anonymous function that fills our chosen hyperparameters in mbtransfer. For example \(x) mbtransfer(x, P = 1, Q = 1) will fit a lag-1 transfer function model on all the random splits.

qvalue

The target False Discovery Rate. Defaults to 0.2.

Examples

data(sim_ts)
w0 <- cbind(sim_ts[[1]]@interventions, matrix(0, nrow = 1, ncol = 3))
w1 <- cbind(sim_ts[[1]]@interventions, matrix(1, nrow = 1, ncol = 3))
select_taxa(sim_ts, w0, w1, \(x) mbtransfer(x, 1, 1, nrounds = 10), n_splits = 1)
#> Training models for split 1/1
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> $taxa
#> $taxa$`1`
#> character(0)
#> 
#> $taxa$`2`
#> character(0)
#> 
#> $taxa$`3`
#> character(0)
#> 
#> $taxa$`4`
#> character(0)
#> 
#> $taxa$`5`
#> character(0)
#> 
#> $taxa$`6`
#> character(0)
#> 
#> $taxa$`7`
#> character(0)
#> 
#> $taxa$`8`
#> character(0)
#> 
#> $taxa$`9`
#> character(0)
#> 
#> $taxa$`10`
#> character(0)
#> 
#> $taxa$`11`
#> character(0)
#> 
#> $taxa$`12`
#> character(0)
#> 
#> $taxa$`13`
#> character(0)
#> 
#> $taxa$`14`
#> character(0)
#> 
#> $taxa$`15`
#> character(0)
#> 
#> $taxa$`16`
#> character(0)
#> 
#> $taxa$`17`
#> character(0)
#> 
#> $taxa$`18`
#> character(0)
#> 
#> $taxa$`19`
#> character(0)
#> 
#> $taxa$`20`
#> character(0)
#> 
#> $taxa$`21`
#> character(0)
#> 
#> $taxa$`22`
#> character(0)
#> 
#> $taxa$`23`
#> character(0)
#> 
#> $taxa$`24`
#> character(0)
#> 
#> $taxa$`25`
#> character(0)
#> 
#> $taxa$`26`
#> character(0)
#> 
#> $taxa$`27`
#> character(0)
#> 
#> $taxa$`28`
#> character(0)
#> 
#> $taxa$`29`
#> character(0)
#> 
#> $taxa$`30`
#> character(0)
#> 
#> 
#> $ms
#> # A tibble: 1,500 × 4
#>        m   lag multisplit taxon
#>    <dbl> <int>      <int> <int>
#>  1     0     1          1     1
#>  2     0     1          1     2
#>  3     0     1          1     3
#>  4     0     1          1     4
#>  5     0     1          1     5
#>  6     0     1          1     6
#>  7     0     1          1     7
#>  8     0     1          1     8
#>  9     0     1          1     9
#> 10     0     1          1    10
#> # ℹ 1,490 more rows
#> 
#> $effects
#> $effects[[1]]
#> , , 1
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 2
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 3
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 4
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 5
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 6
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 7
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 8
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 9
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 10
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 11
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 12
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 13
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 14
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 15
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 16
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 17
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 18
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 19
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 20
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 21
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 22
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 23
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 24
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 25
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 26
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 27
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 28
#> 
#>               [,1]          [,2]
#>  [1,] -2.908556938 -2.591872e+00
#>  [2,]  3.304835853  3.199473e+00
#>  [3,] -1.672057152 -1.255184e+00
#>  [4,] -9.212770394 -1.021007e+01
#>  [5,]  3.307393528  3.146981e+00
#>  [6,]  1.088384628  1.141376e+00
#>  [7,] -1.389186859 -2.142159e+00
#>  [8,]  0.175516129  6.410084e-01
#>  [9,]  0.006427765 -9.767532e-03
#> [10,] -0.218585968 -1.332503e-01
#> [11,]  0.140237502  2.054267e-01
#> [12,] -0.023576736 -2.220230e-01
#> [13,] -0.010618210 -2.038221e-01
#> [14,] -0.332738002 -2.788644e-01
#> [15,] -0.277826309 -2.029644e-02
#> [16,] -0.050538074  4.782881e-01
#> [17,] -0.374387741 -2.554932e-01
#> [18,] -0.816167831  3.031197e-01
#> [19,] -0.771978378 -1.367474e-02
#> [20,]  0.059170746  2.746401e-01
#> [21,] -0.038776398  5.096340e-02
#> [22,] -1.122733071 -6.858864e-01
#> [23,]  0.671307541  7.371428e-02
#> [24,]  0.251770996  9.918213e-05
#> [25,] -0.013117790  2.847757e-01
#> [26,] -0.155976295  5.375767e-02
#> [27,] -0.163640499  4.869122e-01
#> [28,] -0.458649635  2.735691e-01
#> [29,] -0.044634342 -6.692410e-03
#> [30,] -0.578511238 -5.521183e-01
#> [31,] -0.865253448 -4.289627e-01
#> [32,] -0.247351646  2.269838e-02
#> [33,]  0.282413483  2.576504e-01
#> [34,]  0.088269211  4.802578e-01
#> [35,] -0.094549224 -4.219170e-01
#> [36,] -1.598114014 -1.418274e+00
#> [37,] -0.309630394 -6.766546e-06
#> [38,] -0.658884048 -5.328016e-01
#> [39,]  0.160160065 -1.384277e-01
#> [40,]  0.035132408 -5.834064e-01
#> [41,] -0.082344055 -2.889194e-01
#> [42,] -0.651773453 -1.365368e+00
#> [43,]  0.109992027 -3.772259e-02
#> [44,] -0.303059601  1.410036e-01
#> [45,] -0.175096512 -9.040002e-01
#> [46,] -0.027957916  4.708700e-01
#> [47,]  0.400755269  5.613127e-01
#> [48,] -0.198719978 -1.346455e-01
#> [49,] -0.158110596  2.484417e-02
#> [50,] -1.055552732 -2.887249e-01
#> 
#> , , 29
#> 
#>                [,1]          [,2]
#>  [1,]  -3.902304945  -3.571469829
#>  [2,]   3.808153834   3.863841443
#>  [3,]  -2.477584135  -1.829568318
#>  [4,] -10.220128900 -11.306679953
#>  [5,]   3.485311236   3.266839958
#>  [6,]   0.982566130   1.218778542
#>  [7,]  -1.960008349  -2.745772770
#>  [8,]   0.028019269   0.649860632
#>  [9,]  -0.084246159   0.013197081
#> [10,]  -0.402783598  -0.144513505
#> [11,]  -0.028698603   0.114078806
#> [12,]  -0.280524027  -0.643828278
#> [13,]  -0.205164115  -0.402205081
#> [14,]  -0.391655559  -0.369061572
#> [15,]  -0.292265937   0.128278346
#> [16,]   0.229216099   0.735895429
#> [17,]  -0.191776684  -0.275663467
#> [18,]  -1.344562939   0.241282736
#> [19,]  -1.309382757  -0.487945557
#> [20,]  -0.004214582   0.347708577
#> [21,]  -0.240932737   0.316520350
#> [22,]  -1.319295088  -0.779944170
#> [23,]   0.637964816   0.093104726
#> [24,]  -0.042769795  -0.289979821
#> [25,]  -0.157468614  -0.162496612
#> [26,]  -0.317462603   0.001272815
#> [27,]  -0.506915206   0.553859143
#> [28,]  -0.626028152   0.190764314
#> [29,]  -0.172257673  -0.144053141
#> [30,]  -0.729540280  -0.488734813
#> [31,]  -1.088837306  -0.588493915
#> [32,]  -0.184305282   0.079999560
#> [33,]   0.163226128   0.137641112
#> [34,]   0.046179658   0.417019401
#> [35,]  -0.237255959  -0.557350533
#> [36,]  -2.231892450  -1.665715558
#> [37,]  -0.724217755  -0.101173991
#> [38,]  -0.829966704  -0.491165252
#> [39,]   0.074255739  -0.105158193
#> [40,]  -0.585247063  -0.950399649
#> [41,]  -0.162599337  -0.263498942
#> [42,]  -1.387762660  -1.935213134
#> [43,]   0.216536931  -0.038631258
#> [44,]  -0.341080666   0.158435958
#> [45,]  -0.737211840  -1.444645723
#> [46,]  -0.012534573   0.492701281
#> [47,]   0.059820084   0.494723706
#> [48,]  -0.312716938   0.001310349
#> [49,]  -0.217692125   0.022013460
#> [50,]  -1.787649325  -0.999874819
#> 
#> , , 30
#> 
#>               [,1]          [,2]
#>  [1,]  -4.24040195  -3.834235487
#>  [2,]   3.76636871   3.913480327
#>  [3,]  -2.82177112  -2.031104338
#>  [4,] -10.58836104 -11.649988878
#>  [5,]   3.34327482   3.177285194
#>  [6,]   0.83909850   1.171793870
#>  [7,]  -2.19249841  -2.973706382
#>  [8,]  -0.14448302   0.539342948
#>  [9,]  -0.23352614  -0.055846010
#> [10,]  -0.58335898  -0.233026913
#> [11,]  -0.18453542   0.023978631
#> [12,]  -0.49490985  -0.842693420
#> [13,]  -0.39353207  -0.507275536
#> [14,]  -0.52549592  -0.446851390
#> [15,]  -0.46124579   0.039501803
#> [16,]   0.16757424   0.712155240
#> [17,]  -0.27459615  -0.390076910
#> [18,]  -1.63941016   0.136733509
#> [19,]  -1.53701376  -0.666885149
#> [20,]  -0.15231335   0.270521675
#> [21,]  -0.42289375   0.301637559
#> [22,]  -1.49931744  -0.876565979
#> [23,]   0.45573151   0.014569896
#> [24,]  -0.24366118  -0.439838932
#> [25,]  -0.32370336  -0.306566738
#> [26,]  -0.48816633  -0.091123218
#> [27,]  -0.66243104   0.534410863
#> [28,]  -0.76003888   0.100282919
#> [29,]  -0.27117914  -0.205190397
#> [30,]  -0.87805557  -0.581454027
#> [31,]  -1.26367206  -0.720536663
#> [32,]  -0.28893868   0.007355758
#> [33,]  -0.04774866   0.048711504
#> [34,]  -0.07675627   0.335536389
#> [35,]  -0.37657615  -0.652230706
#> [36,]  -2.49835282  -1.813933668
#> [37,]  -0.93201964  -0.237671012
#> [38,]  -1.02423046  -0.569957529
#> [39,]  -0.13606046  -0.206294264
#> [40,]  -0.89676067  -1.190375396
#> [41,]  -0.37512579  -0.367263249
#> [42,]  -1.67734137  -2.084311667
#> [43,]   0.08763929  -0.144107478
#> [44,]  -0.48355411   0.074687594
#> [45,]  -1.01439437  -1.627338886
#> [46,]  -0.16157859   0.383137794
#> [47,]  -0.18297422   0.405518078
#> [48,]  -0.48480070  -0.069180897
#> [49,]  -0.36155308  -0.067707743
#> [50,]  -2.05678637  -1.219168107
#> 
#> 
#>