Significant Taxa using Mirror Statistics

This selects taxa through a data splitting procedure. It retrains an mbtransfer model across random splits of the data. Partial dependence profiles for pairs of splits are compared with one another -- if they agree for a given feature, then that feature is considered more likely to have a true effect on the response. This function only supports inference of the effects of interventions on taxa responses, but the same principle could apply to estimate significant relationships between taxa.

Usage

select_taxa(ts, w0, w1, tr_fun, qvalue = 0.2, ...)

Arguments

ts

An object of class ts_inter containing the time-varying microbiome community, environmental interventions, and static host features data. The columns for each element of the values matrix are expected to be consecutive timepoints in an changing community. ts_from_dfs() is a helper to create these objects from data.frames with the relevant information.

w0

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

w1

One of the counterfactuals with which to compute partial dependence profiles. See steps or pulses for helpers in generating these counterfactuals. The procedure concatenates these counterfactuals to the end of the series and computes the difference in the forecasts.

tr_fun

A function that can be used to train over random splits. In the examples in this package, we use an anonymous function that fills our chosen hyperparameters in mbtransfer. For example \(x) mbtransfer(x, P = 1, Q = 1) will fit a lag-1 transfer function model on all the random splits.

qvalue

The target False Discovery Rate. Defaults to 0.2.

Examples

data(sim_ts)
w0 <- cbind(sim_ts[[1]]@interventions, matrix(0, nrow = 1, ncol = 3))
w1 <- cbind(sim_ts[[1]]@interventions, matrix(1, nrow = 1, ncol = 3))
select_taxa(sim_ts, w0, w1, \(x) mbtransfer(x, 1, 1, nrounds = 10), n_splits = 1)
#> Training models for split 1/1
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> $taxa
#> $taxa$`1`
#> character(0)
#> 
#> $taxa$`2`
#> character(0)
#> 
#> $taxa$`3`
#> character(0)
#> 
#> $taxa$`4`
#> character(0)
#> 
#> $taxa$`5`
#> character(0)
#> 
#> $taxa$`6`
#> character(0)
#> 
#> $taxa$`7`
#> character(0)
#> 
#> $taxa$`8`
#> character(0)
#> 
#> $taxa$`9`
#> character(0)
#> 
#> $taxa$`10`
#> character(0)
#> 
#> $taxa$`11`
#> character(0)
#> 
#> $taxa$`12`
#> character(0)
#> 
#> $taxa$`13`
#> character(0)
#> 
#> $taxa$`14`
#> character(0)
#> 
#> $taxa$`15`
#> character(0)
#> 
#> $taxa$`16`
#> character(0)
#> 
#> $taxa$`17`
#> character(0)
#> 
#> $taxa$`18`
#> character(0)
#> 
#> $taxa$`19`
#> character(0)
#> 
#> $taxa$`20`
#> character(0)
#> 
#> $taxa$`21`
#> character(0)
#> 
#> $taxa$`22`
#> character(0)
#> 
#> $taxa$`23`
#> character(0)
#> 
#> $taxa$`24`
#> character(0)
#> 
#> $taxa$`25`
#> character(0)
#> 
#> $taxa$`26`
#> character(0)
#> 
#> $taxa$`27`
#> character(0)
#> 
#> $taxa$`28`
#> character(0)
#> 
#> $taxa$`29`
#> character(0)
#> 
#> $taxa$`30`
#> character(0)
#> 
#> 
#> $ms
#> # A tibble: 1,500 × 4
#>        m   lag multisplit taxon
#>    <dbl> <int>      <int> <int>
#>  1     0     1          1     1
#>  2     0     1          1     2
#>  3     0     1          1     3
#>  4     0     1          1     4
#>  5     0     1          1     5
#>  6     0     1          1     6
#>  7     0     1          1     7
#>  8     0     1          1     8
#>  9     0     1          1     9
#> 10     0     1          1    10
#> # ℹ 1,490 more rows
#> 
#> $effects
#> $effects[[1]]
#> , , 1
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 2
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 3
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 4
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 5
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 6
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 7
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 8
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 9
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 10
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 11
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 12
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 13
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 14
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 15
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 16
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 17
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 18
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 19
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 20
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 21
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 22
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 23
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 24
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 25
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 26
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 27
#> 
#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]    0    0
#>  [3,]    0    0
#>  [4,]    0    0
#>  [5,]    0    0
#>  [6,]    0    0
#>  [7,]    0    0
#>  [8,]    0    0
#>  [9,]    0    0
#> [10,]    0    0
#> [11,]    0    0
#> [12,]    0    0
#> [13,]    0    0
#> [14,]    0    0
#> [15,]    0    0
#> [16,]    0    0
#> [17,]    0    0
#> [18,]    0    0
#> [19,]    0    0
#> [20,]    0    0
#> [21,]    0    0
#> [22,]    0    0
#> [23,]    0    0
#> [24,]    0    0
#> [25,]    0    0
#> [26,]    0    0
#> [27,]    0    0
#> [28,]    0    0
#> [29,]    0    0
#> [30,]    0    0
#> [31,]    0    0
#> [32,]    0    0
#> [33,]    0    0
#> [34,]    0    0
#> [35,]    0    0
#> [36,]    0    0
#> [37,]    0    0
#> [38,]    0    0
#> [39,]    0    0
#> [40,]    0    0
#> [41,]    0    0
#> [42,]    0    0
#> [43,]    0    0
#> [44,]    0    0
#> [45,]    0    0
#> [46,]    0    0
#> [47,]    0    0
#> [48,]    0    0
#> [49,]    0    0
#> [50,]    0    0
#> 
#> , , 28
#> 
#>              [,1]          [,2]
#>  [1,] -2.98653412  -2.553578377
#>  [2,]  3.33928949   3.204713583
#>  [3,] -1.61930425  -1.296922684
#>  [4,] -7.12169877 -12.312889076
#>  [5,]  3.36669827   3.100467023
#>  [6,]  1.26352393   0.966639519
#>  [7,] -1.93695450  -1.524941399
#>  [8,]  0.52460480   0.281187057
#>  [9,]  0.19446754  -0.228019714
#> [10,] -0.11238482  -0.216789143
#> [11,]  0.27329871   0.083112240
#> [12,]  0.13154221  -0.349139214
#> [13,] -0.16778278  -0.017475128
#> [14,] -0.59704485  -0.017874718
#> [15,] -0.15507317  -0.088931106
#> [16,]  0.19571590   0.219628879
#> [17,] -0.18901253  -0.448207855
#> [18,] -0.51951113  -0.011233330
#> [19,] -0.43588638  -0.260960579
#> [20,]  0.25041008   0.100056648
#> [21,]  0.09119415  -0.066288948
#> [22,] -1.07272062  -0.769369125
#> [23,]  0.57747841   0.247939133
#> [24,]  0.03825951   0.204178810
#> [25,]  0.38737392  -0.169234253
#> [26,]  0.05336571  -0.142472267
#> [27,]  0.39178796  -0.090422131
#> [28,] -0.29771709   0.106037140
#> [29,]  0.06035995  -0.132138729
#> [30,] -0.55227852  -0.577237129
#> [31,] -0.25280762  -1.021163940
#> [32,] -0.11456776  -0.058036804
#> [33,]  0.46944046   0.112959862
#> [34,]  0.49189472   0.053747154
#> [35,] -0.30817795  -0.224709965
#> [36,] -1.41098118  -1.624135971
#> [37,]  0.20257160  -0.529217788
#> [38,] -0.39295292  -0.733314514
#> [39,] -0.17782021   0.209227607
#> [40,]  0.05769920  -0.599467346
#> [41,]  0.05662346  -0.481435776
#> [42,] -0.48604679  -1.561240083
#> [43,]  0.17309761  -0.015521049
#> [44,]  0.01285172  -0.157241821
#> [45,] -0.50444603  -0.576297760
#> [46,]  0.42437935  -0.009949684
#> [47,]  0.60873506   0.377931595
#> [48,] -0.51931858   0.187018394
#> [49,]  0.09988022  -0.216713905
#> [50,] -0.81845432  -0.434985195
#> 
#> , , 29
#> 
#>              [,1]          [,2]
#>  [1,] -3.88641010  -3.466198490
#>  [2,]  3.98781552   3.665862447
#>  [3,] -2.13505675  -2.133792128
#>  [4,] -8.00747601 -13.434502261
#>  [5,]  3.59681913   3.158486548
#>  [6,]  1.21174743   0.987301633
#>  [7,] -2.64789009  -2.002400262
#>  [8,]  0.45954956   0.207407157
#>  [9,]  0.25622368  -0.405484358
#> [10,] -0.26019254  -0.285470293
#> [11,]  0.12204601  -0.046332461
#> [12,] -0.07106204  -0.815780004
#> [13,] -0.27277794  -0.289021787
#> [14,] -0.75634887  -0.025384733
#> [15,] -0.11304964  -0.026053406
#> [16,]  0.47483901   0.502103590
#> [17,] -0.11002659  -0.376926854
#> [18,] -0.63982044  -0.489212195
#> [19,] -0.89837730  -0.863778410
#> [20,]  0.31445444  -0.014098111
#> [21,]  0.26156248  -0.183394432
#> [22,] -1.25405639  -0.869193963
#> [23,]  0.87745769  -0.090465409
#> [24,] -0.13121637  -0.107892150
#> [25,]  0.04139396  -0.370417300
#> [26,]  0.08917972  -0.341218767
#> [27,]  0.30123759  -0.238357385
#> [28,] -0.44437604  -0.004282702
#> [29,] -0.16603142  -0.228851182
#> [30,] -0.50687229  -0.710920902
#> [31,] -0.46963667  -1.183224837
#> [32,] -0.09670707   0.078921341
#> [33,]  0.27259652   0.070715178
#> [34,]  0.55968459  -0.085126423
#> [35,] -0.41001063  -0.387692440
#> [36,] -1.79900882  -2.115385510
#> [37,] -0.17943444  -0.701501188
#> [38,] -0.40915201  -0.842033999
#> [39,] -0.30494236   0.236254261
#> [40,] -0.45948072  -1.099168550
#> [41,]  0.12353363  -0.599117869
#> [42,] -1.05092244  -2.260776997
#> [43,]  0.40695629  -0.136987141
#> [44,] -0.09259324  -0.217320329
#> [45,] -0.80319230  -1.396713688
#> [46,]  0.55048915  -0.146227950
#> [47,]  0.47109999   0.150471937
#> [48,] -0.54526897   0.284105959
#> [49,]  0.07069054  -0.212399642
#> [50,] -1.48741039  -1.310167903
#> 
#> , , 30
#> 
#>               [,1]         [,2]
#>  [1,] -4.139637652  -3.79834990
#>  [2,]  4.029723077   3.63326286
#>  [3,] -2.357613246  -2.43404202
#>  [4,] -8.296770482 -13.84277148
#>  [5,]  3.503350780   2.99434971
#>  [6,]  1.109363590   0.88601251
#>  [7,] -2.876019455  -2.23593746
#>  [8,]  0.332913376   0.04347885
#>  [9,]  0.174906004  -0.56194469
#> [10,] -0.371130716  -0.45057167
#> [11,] -0.002518484  -0.17977559
#> [12,] -0.209567388  -1.04377038
#> [13,] -0.392374833  -0.47854687
#> [14,] -0.842907837  -0.16184383
#> [15,] -0.217306251  -0.17818138
#> [16,]  0.458168518   0.45488871
#> [17,] -0.184550126  -0.51071669
#> [18,] -0.812586058  -0.73448531
#> [19,] -1.099003951  -1.10275094
#> [20,]  0.235639390  -0.16319967
#> [21,]  0.194528375  -0.30836871
#> [22,] -1.396509466  -1.03687191
#> [23,]  0.819845449  -0.31445197
#> [24,] -0.268838905  -0.32377470
#> [25,] -0.129639330  -0.52518967
#> [26,] -0.001451084  -0.50367557
#> [27,]  0.214403618  -0.34955720
#> [28,] -0.553525221  -0.13775333
#> [29,] -0.247669992  -0.30418273
#> [30,] -0.589905126  -0.87377083
#> [31,] -0.615747792  -1.34529695
#> [32,] -0.170668988  -0.01902430
#> [33,]  0.107100328  -0.07272023
#> [34,]  0.477247726  -0.21649121
#> [35,] -0.499802862  -0.53706369
#> [36,] -1.975930078  -2.35871851
#> [37,] -0.343639624  -0.90493765
#> [38,] -0.513018631  -0.97267255
#> [39,] -0.460255555   0.08450015
#> [40,] -0.704916046  -1.38370671
#> [41,] -0.009344623  -0.81961838
#> [42,] -1.252309663  -2.51733194
#> [43,]  0.345897379  -0.28729387
#> [44,] -0.206164655  -0.33296992
#> [45,] -0.998247419  -1.66498820
#> [46,]  0.479573568  -0.32339146
#> [47,]  0.340841906  -0.05742782
#> [48,] -0.666474456   0.18537985
#> [49,] -0.036192712  -0.33539727
#> [50,] -1.714465822  -1.55188614
#> 
#> 
#>