Build a ts

Build a ts

Usage

ts_from_dfs(reads, interventions, metadata, subject_data = NULL)

Examples

library(readr)
library(tibble)
subject <- read_csv("https://figshare.com/ndownloader/files/40275934/subject.csv")
#> Rows: 20 Columns: 2
#> -- Column specification --------------------------------------------------------
#> Delimiter: ","
#> chr (2): subject, diet
#> 
#> i Use `spec()` to retrieve the full column specification for this data.
#> i Specify the column types or set `show_col_types = FALSE` to quiet this message.
interventions <- read_csv("https://figshare.com/ndownloader/files/40279171/interventions.csv") |>
column_to_rownames("sample")
#> Rows: 236 Columns: 3
#> -- Column specification --------------------------------------------------------
#> Delimiter: ","
#> chr (1): sample
#> dbl (2): Plant, Animal
#> 
#> i Use `spec()` to retrieve the full column specification for this data.
#> i Specify the column types or set `show_col_types = FALSE` to quiet this message.
reads <- read_csv("https://figshare.com/ndownloader/files/40279108/reads.csv") |>
 column_to_rownames("sample")
#> Rows: 236 Columns: 192
#> -- Column specification --------------------------------------------------------
#> Delimiter: ","
#> chr   (1): sample
#> dbl (191): Otu000001, Otu000002, Otu000003, Otu000004, Otu000005, Otu000006,...
#> 
#> i Use `spec()` to retrieve the full column specification for this data.
#> i Specify the column types or set `show_col_types = FALSE` to quiet this message.
samples <- read_csv("https://figshare.com/ndownloader/files/40275943/samples.csv")
#> Rows: 236 Columns: 4
#> -- Column specification --------------------------------------------------------
#> Delimiter: ","
#> chr (3): sample, subject, condition
#> dbl (1): time
#> 
#> i Use `spec()` to retrieve the full column specification for this data.
#> i Specify the column types or set `show_col_types = FALSE` to quiet this message.
ts <- as.matrix(reads) |>
  ts_from_dfs(interventions, samples, subject)
#> Joining with `by = join_by(subject)`
ts
#> # A ts_inter object | 191 taxa | 20 subjects | 11.80 <U+00B1> 3.57 timepoints:
#> 
#> Plant5:
#>                       DD53            DD111             DD24             DD29
#> Otu000001 10.2182984901644 9.14941182693663 8.39212694793041 9.39899720494429
#> Otu000002                0                0 5.15032449125274 8.45355572278925
#> Otu000003 6.41298639815534                0  3.5089592334818 3.56986494940344
#> Otu000004 7.45977126049619 6.99747111345779 6.85504008284069  8.0400754435393
#>                   <U+22EE>         <U+22EE>         <U+22EE>         <U+22EE>
#>                   
#> Otu000001 <U+2026>
#> Otu000002 <U+2026>
#> Otu000003 <U+2026>
#> Otu000004 <U+2026>
#>           <U+22F1>
#> 
#> Plant7:
#>                       DD79             DD34             DD72            DD144
#> Otu000001 9.31064796332915 9.36846144577821 8.05207430379548 9.73394258993618
#> Otu000002                0                0 4.83763438231674 1.84428530870696
#> Otu000003 7.20133580603692 7.57511920224131 5.67443724637428 7.12722792269608
#> Otu000004  9.3751864834759 9.59128858883673 8.31292026544233 9.43524314853652
#>                   <U+22EE>         <U+22EE>         <U+22EE>         <U+22EE>
#>                   
#> Otu000001 <U+2026>
#> Otu000002 <U+2026>
#> Otu000003 <U+2026>
#> Otu000004 <U+2026>
#>           <U+22F1>
#> 
#> Plant4:
#>                       DD31            DD146             DD38            DD138
#> Otu000001 8.17627997254723 8.51162759090965 8.28399997223762 8.46589312749772
#> Otu000002 4.54496581622368                0                0                0
#> Otu000003 8.47143621467469  8.2262417618481 7.68084465953917 6.52660031646268
#> Otu000004                0                0                0                0
#>                   <U+22EE>         <U+22EE>         <U+22EE>         <U+22EE>
#>                   
#> Otu000001 <U+2026>
#> Otu000002 <U+2026>
#> Otu000003 <U+2026>
#> Otu000004 <U+2026>
#>           <U+22F1>
#> 
#> and 17 more subjects.